--- EventManager.cpp
@@ -1,183 +1,0 @@
-//
-// COMP 371 Assignment Framework
-//
-// Created by Nicolas Bergeron on 8/7/14.
-// Updated by Gary Chang on 14/1/15
-//
-// Copyright (c) 2014-2019 Concordia University. All rights reserved.
-//
-
-#include "EventManager.h"
-#include "Renderer.h"
-
-#include <GLFW/glfw3.h>
-
-#include <stdio.h>
-#include <iostream>
-#include <fstream>
-
-#include <stdlib.h>
-#include <stdio.h>
-#include <time.h>
-
-using namespace std;
-
-
-// Time
-double EventManager::sLastFrameTime = glfwGetTime();
-float  EventManager::sFrameTime = 0.0f;
-
-// Mouse
-double EventManager::sLastMousePositionX = 0.0f;
-float  EventManager::sMouseDeltaX = 0.0f;
-double EventManager::sLastMousePositionY = 0.0f;
-float  EventManager::sMouseDeltaY = 0.0f;
-
-// Window
-GLFWwindow* EventManager::spWindow = nullptr;
-float EventManager::window_height = 768;
-float EventManager::window_width = 1024;
-
-
-void EventManager::Initialize()
-{
-	// Initialise GLFW
-	if( !glfwInit() )
-	{
-		fprintf( stderr, "Failed to initialize GLFW\n" );
-		exit(-1);
-	}
-		
-	glfwWindowHint(GLFW_SAMPLES, 4);
-    
-#if defined(PLATFORM_OSX)
-    // OS X would use legacy opengl by default, and wouldn't be able to load shaders
-    // This is the proper way to setup GLFW to use modern OpenGL
-    glfwWindowHint(GLFW_CONTEXT_VERSION_MAJOR, 3);
-    glfwWindowHint(GLFW_CONTEXT_VERSION_MINOR, 2);
-    glfwWindowHint(GLFW_OPENGL_PROFILE, GLFW_OPENGL_CORE_PROFILE);
-    glfwWindowHint(GLFW_OPENGL_FORWARD_COMPAT, GL_TRUE);
-	glfwWindowHint(GLFW_DEPTH_BITS, 32);
-    
-#else
-    // Allow older laptops to run the framework, even though, our framework
-    // is compatible with OpenGL 3.3 (which removed many deprecated features)
-	glfwWindowHint(GLFW_CONTEXT_VERSION_MAJOR, 2);
-	glfwWindowHint(GLFW_CONTEXT_VERSION_MINOR, 1);
-	glfwWindowHint(GLFW_DEPTH_BITS, 32);
-#endif
-    
-    
-	// Open a window and create its OpenGL context
-	glfwWindowHint(GLFW_RESIZABLE, 0);
-	const GLFWvidmode * mode = glfwGetVideoMode(glfwGetPrimaryMonitor());
-	window_width = mode->width;
-	window_height = mode->height;
-	
-	/*int count;
-	GLFWmonitor** monitors = glfwGetMonitors(&count);*/
-	spWindow = glfwCreateWindow(1024, 768, "Vaporwave", nullptr, nullptr);
-	//spWindow = glfwCreateWindow(window_width, window_height, "Vaporwave", glfwGetPrimaryMonitor(), nullptr);
-
-	if (spWindow == nullptr)
-	{
-		fprintf( stderr, "Failed to open GLFW window. If you have an Intel GPU, they are not 3.3 compatible. Try the 2.1 version of the tutorials.\n" );
-		glfwTerminate();
-		exit(-1);
-	}
-
-	// Ensure we can capture the escape key being pressed below
-	glfwSetInputMode(spWindow, GLFW_STICKY_KEYS, GL_TRUE);
-
-	// Initial mouse position
-	glfwPollEvents();
-	double x, y;
-	glfwGetCursorPos(spWindow, &x, &y);
-
-	sLastMousePositionX = x;
-	sLastMousePositionY = y;
-
-	// Initial time
-	sLastFrameTime = glfwGetTime();
-    srand((unsigned int) time(nullptr));
-}
-
-void EventManager::Shutdown()
-{
-	// Close OpenGL window and terminate GLFW
-	glfwTerminate();
-	spWindow = nullptr;
-}
-
-void EventManager::Update()
-{
-	// Update inputs / events
-	glfwPollEvents();
-
-	// Update mouse position
-	double x, y;
-	glfwGetCursorPos(spWindow, &x, &y);
-	sMouseDeltaX = static_cast<float>( x - sLastMousePositionX );
-	sMouseDeltaY = static_cast<float>( y - sLastMousePositionY );
-	sLastMousePositionX = x;
-	sLastMousePositionY = y;
-
-	// Update frame time
-	double currentTime = glfwGetTime();
-	sFrameTime = static_cast<float>(currentTime - sLastFrameTime);
-	sLastFrameTime = currentTime;
-}
-
-float EventManager::GetFrameTime()
-{
-	return sFrameTime;
-}
-
-bool EventManager::ExitRequested()
-{
-	return glfwGetKey(spWindow, GLFW_KEY_ESCAPE ) == GLFW_PRESS || glfwWindowShouldClose(spWindow);
-}
-
-GLFWwindow* EventManager::GetWindow()
-{
-	return spWindow;
-}
-
-float EventManager::GetMouseMotionX()
-{
-	return sMouseDeltaX;
-}
-
-float EventManager::GetMouseMotionY()
-{
-	return sMouseDeltaY;
-}
-
-void EventManager::EnableMouseCursor()
-{
-	glfwSetInputMode(spWindow, GLFW_CURSOR, GLFW_CURSOR_NORMAL);
-}
-
-void EventManager::DisableMouseCursor()
-{
-	glfwSetInputMode(spWindow, GLFW_CURSOR, GLFW_CURSOR_DISABLED);
-}
-
-float EventManager::GetRandomFloat(float min, float max)
-{
-    float value = static_cast <float> (rand()) / static_cast <float> (RAND_MAX);
-
-    return min + value*(max - min);
-}
-
-int EventManager::GetRandomInt(int min, int max)
-{
-	int value = static_cast <int> (rand()) / static_cast <int> (RAND_MAX);
-
-	return min + value * (max - min);
-}
-
-int EventManager::GetRandomInt()
-{
-	return static_cast <int> (rand());
-}
