--- EventManager.h
@@ -1,55 +1,0 @@
-//
-// COMP 371 Assignment Framework
-//
-// Created by Nicolas Bergeron on 8/7/14.
-// Updated by Gary Chang on 14/1/15
-//
-// Copyright (c) 2014-2019 Concordia University. All rights reserved.
-//
-
-#pragma once
-
-struct GLFWwindow;
-
-class EventManager
-{
-public:
-	static void Initialize();
-	static void Shutdown();
-
-	static void Update();
-
-	static float GetFrameTime();
-	static bool ExitRequested();
-
-	static GLFWwindow* GetWindow();
-
-	static float GetMouseMotionX();
-	static float GetMouseMotionY();
-
-	static void EnableMouseCursor();
-	static void DisableMouseCursor();
-    static float GetRandomFloat(float min, float max);
-	static int GetRandomInt(int min, int max);
-	static int GetRandomInt();
-
-	static float GetWindowHight() { return window_height; }
-	static float GetWindowWidth() { return window_width; }
-
-private:
-	// Time
-	static double sLastFrameTime;
-	static float sFrameTime;
-
-	// Mouse
-	static double sLastMousePositionX;
-	static float  sMouseDeltaX;
-	static double sLastMousePositionY;
-	static float  sMouseDeltaY;
-
-	static float window_width;
-	static float window_height;
-
-	// Window
-	static GLFWwindow* spWindow;
-};
