--- Model.cpp
@@ -1,202 +1,0 @@
-//
-// COMP 371 Assignment Framework
-//
-// Created by Nicolas Bergeron on 8/7/14.
-// Updated by Gary Chang on 14/1/15
-//
-// Copyright (c) 2014-2019 Concordia University. All rights reserved.
-//
-
-#include "Model.h"
-#include "Animation.h"
-#include "World.h"
-#include "WorldBlock.h"
-#include "ParticleEmitter.h"
-#include "ParticleDescriptor.h"
-#include "ParticleSystem.h"
-#include "Buildings.h"
-#include <glm/gtc/matrix_transform.hpp>
-#include <glm/common.hpp>
-
-using namespace std;
-using namespace glm;
-
-Model::Model() 
-: mName("UNNAMED"), mPosition(0.0f, 0.0f, 0.0f), mScaling(1.0f, 1.0f, 1.0f), mRotationAxis(0.0f, 1.0f, 0.0f),
-  mRotationAngleInDegrees(0.0f), mAnimation(nullptr)
-{
-}
-
-Model::~Model()
-{
-}
-
-void Model::Update(float dt)
-{
-	mat4 worldMatrix = GetWorldMatrix();
-	mPosition = vec3(worldMatrix[3][0], worldMatrix[3][1], worldMatrix[3][2]);
-}
-
-void Model::Draw(mat4 offsetMatrix)
-{
-
-}
-
-
-void Model::Load(ci_istringstream& iss)
-{
-	ci_string line;
-
-	// Parse model line by line
-	while(std::getline(iss, line))
-	{
-		// Splitting line into tokens
-		ci_istringstream strstr(line);
-		istream_iterator<ci_string, char, ci_char_traits> it(strstr);
-		istream_iterator<ci_string, char, ci_char_traits> end;
-		vector<ci_string> token(it, end);
-
-		if (ParseLine(token) == false)
-		{
-			fprintf(stderr, "Error loading scene file... token:  %s!", token[0].c_str());
-			getchar();
-			exit(-1);
-		}
-	}
-}
-
-bool Model::ParseLine(const std::vector<ci_string> &token)
-{
-    if (token.empty() == false)
-	{
-		if (token[0].empty() == false && token[0][0] == '#')
-		{
-			return true;
-		}
-		else if (token[0] == "name")
-		{
-			assert(token.size() > 2);
-			assert(token[1] == "=");
-
-			mName = token[2];	
-		}
-		else if (token[0] == "position")
-		{
-			assert(token.size() > 4);
-			assert(token[1] == "=");
-
-			mPosition.x = static_cast<float>(atof(token[2].c_str()));
-			mPosition.y = static_cast<float>(atof(token[3].c_str()));
-			mPosition.z = static_cast<float>(atof(token[4].c_str()));
-		}
-		else if (token[0] == "rotation")
-		{
-			assert(token.size() > 4);
-			assert(token[1] == "=");
-
-			mRotationAxis.x = static_cast<float>(atof(token[2].c_str()));
-			mRotationAxis.y = static_cast<float>(atof(token[3].c_str()));
-			mRotationAxis.z = static_cast<float>(atof(token[4].c_str()));
-			mRotationAngleInDegrees = static_cast<float>(atof(token[5].c_str()));
-
-			glm::normalize(mRotationAxis);
-		}
-		else if (token[0] == "scaling")
-		{
-			assert(token.size() > 4);
-			assert(token[1] == "=");
-
-			mScaling.x = static_cast<float>(atof(token[2].c_str()));
-			mScaling.y = static_cast<float>(atof(token[3].c_str()));
-			mScaling.z = static_cast<float>(atof(token[4].c_str()));
-
-			ci_string temp = GetName();
-			if (GetName() == "\"Building\"")
-				Buildings::setBuildingDefaultSize(mScaling);
-		}
-		else if (token[0] == "animation")
-		{
-			assert(token.size() > 2);
-			assert(token[1] == "=");
-
-			ci_string animName = token[2];
-            
-            //mAnimation = World::getWorldInstance()->getWorldBlock()->FindAnimation(animName);
-			mAnimation = World::getWorldInstance()->FindAnimation(animName);
-		}
-        else if (token[0] == "particleemitter")
-        {
-            assert(token.size() > 2);
-            assert(token[1] == "=");
-
-            ParticleDescriptor* desc = World::getWorldInstance()->FindParticleDescriptor(token[2]);
-            assert(desc != nullptr);
-            
-            ParticleEmitter* emitter = new ParticleEmitter(vec3(0.0f, 0.0f, 0.0f), this);
-            
-            ParticleSystem* ps = new ParticleSystem(emitter, desc);
-            //World::getWorldInstance()->getWorldBlock()->AddParticleSystem(ps);
-			if (GetName() == "\"MainCharacter\"")
-				World::getWorldInstance()->AddMCParticleSystem(ps);
-			else
-				World::getWorldInstance()->AddParticleSystem(ps);
-        }
-		else if (token[0] == "properties")
-		{
-			assert(token.size() > 5);
-			assert(token[1] == "=");
-
-			properties.x = static_cast<float>(atof(token[2].c_str()));
-			properties.y = static_cast<float>(atof(token[3].c_str()));
-			properties.z = static_cast<float>(atof(token[4].c_str()));
-			properties.w = static_cast<float>(atof(token[5].c_str()));
-		}
-		else
-		{
-			return false;
-		}
-	}
-
-	return true;
-}
-
-glm::mat4 Model::GetWorldMatrix() const
-{
-	// @TODO 2 - You must build the world matrix from the position, scaling and rotation informations
-    //           If the model has an animation, get the world transform from the animation.
-	mat4 worldMatrix;
-
-    // Solution TRS
-#if 1
-    if (mAnimation)
-    {
-        // Get world transform from animation key frames / current time
-        worldMatrix = mAnimation->GetAnimationWorldMatrix();
-    }
-    else
-    {
-        mat4 t = glm::translate(mat4(1.0f), mPosition);
-        mat4 r = glm::rotate(mat4(1.0f), glm::radians(mRotationAngleInDegrees), mRotationAxis);
-        mat4 s = glm::scale(mat4(1.0f), mScaling);
-        worldMatrix = t * r * s;
-    }
-#endif
-    
-	return worldMatrix;
-}
-
-void Model::SetPosition(glm::vec3 position)
-{
-	mPosition = position;
-}
-
-void Model::SetScaling(glm::vec3 scaling)
-{
-	mScaling = scaling;
-}
-
-void Model::SetRotation(glm::vec3 axis, float angleDegrees)
-{
-	mRotationAxis = axis;
-	mRotationAngleInDegrees = angleDegrees;
-}
