--- SphereModel.cpp
@@ -1,1367 +1,0 @@
-//
-// COMP 371 Assignment Framework
-//
-// Created by Nicolas Bergeron on 8/7/14.
-// Updated by Gary Chang on 14/1/15
-//
-// Copyright (c) 2014-2019 Concordia University. All rights reserved.
-//
-
-#include "SphereModel.h"
-#include "Renderer.h"
-
-using namespace glm;
-
-SphereModel::SphereModel(vec3 size) : Model()
-{
-    Vertex vertexBuffer[] = {
-        // position,                                    normal,                              color
-        { vec3(0.000000, 0.000000, -1.000000), vec3(0.000000, 0.000000, -1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.173648, 0.000000, -0.984808), vec3(0.173648, 0.000000, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, 0.030154, -0.984808), vec3(0.171010, 0.030154, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.342020, 0.000000, -0.939693), vec3(0.342020, 0.000000, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.336824, 0.059391, -0.939693), vec3(0.336824, 0.059391, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.500000, 0.000000, -0.866025), vec3(0.500000, 0.000000, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, 0.086824, -0.866025), vec3(0.492404, 0.086824, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.642788, 0.000000, -0.766044), vec3(0.642788, 0.000000, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.633022, 0.111619, -0.766044), vec3(0.633022, 0.111619, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.766044, 0.000000, -0.642788), vec3(0.766044, 0.000000, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.754407, 0.133022, -0.642788), vec3(0.754407, 0.133022, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.866025, 0.000000, -0.500000), vec3(0.866025, 0.000000, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.852869, 0.150384, -0.500000), vec3(0.852869, 0.150384, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.939693, 0.000000, -0.342020), vec3(0.939693, 0.000000, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.925417, 0.163176, -0.342020), vec3(0.925417, 0.163176, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.984808, 0.000000, -0.173648), vec3(0.984808, 0.000000, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.969846, 0.171010, -0.173648), vec3(0.969846, 0.171010, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(1.000000, 0.000000, 0.000000), vec3(1.000000, 0.000000, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.984808, 0.173648, 0.000000), vec3(0.984808, 0.173648, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.984808, 0.000000, 0.173648), vec3(0.984808, 0.000000, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.969846, 0.171010, 0.173648), vec3(0.969846, 0.171010, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.939693, 0.000000, 0.342020), vec3(0.939693, 0.000000, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.925417, 0.163176, 0.342020), vec3(0.925417, 0.163176, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.866025, 0.000000, 0.500000), vec3(0.866025, 0.000000, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.852869, 0.150384, 0.500000), vec3(0.852869, 0.150384, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.766044, 0.000000, 0.642788), vec3(0.766044, 0.000000, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.754407, 0.133022, 0.642788), vec3(0.754407, 0.133022, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.642788, 0.000000, 0.766044), vec3(0.642788, 0.000000, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.633022, 0.111619, 0.766044), vec3(0.633022, 0.111619, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.500000, 0.000000, 0.866025), vec3(0.500000, 0.000000, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, 0.086824, 0.866025), vec3(0.492404, 0.086824, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.342020, 0.000000, 0.939693), vec3(0.342020, 0.000000, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.336824, 0.059391, 0.939693), vec3(0.336824, 0.059391, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.173648, 0.000000, 0.984808), vec3(0.173648, 0.000000, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, 0.030154, 0.984808), vec3(0.171010, 0.030154, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, 1.000000), vec3(0.000000, 0.000000, 1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, 0.030154, 0.984808), vec3(0.171010, 0.030154, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.163176, 0.059391, 0.984808), vec3(0.163176, 0.059391, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.336824, 0.059391, 0.939693), vec3(0.336824, 0.059391, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, 0.116978, 0.939693), vec3(0.321394, 0.116978, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, 0.086824, 0.866025), vec3(0.492404, 0.086824, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.469846, 0.171010, 0.866025), vec3(0.469846, 0.171010, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.633022, 0.111619, 0.766044), vec3(0.633022, 0.111619, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.604023, 0.219846, 0.766044), vec3(0.604023, 0.219846, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.754407, 0.133022, 0.642788), vec3(0.754407, 0.133022, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.719846, 0.262003, 0.642788), vec3(0.719846, 0.262003, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.852869, 0.150384, 0.500000), vec3(0.852869, 0.150384, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.813798, 0.296198, 0.500000), vec3(0.813798, 0.296198, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.925417, 0.163176, 0.342020), vec3(0.925417, 0.163176, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.883022, 0.321394, 0.342020), vec3(0.883022, 0.321394, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.969846, 0.171010, 0.173648), vec3(0.969846, 0.171010, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.925417, 0.336824, 0.173648), vec3(0.925417, 0.336824, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.984808, 0.173648, 0.000000), vec3(0.984808, 0.173648, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.939693, 0.342020, 0.000000), vec3(0.939693, 0.342020, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.969846, 0.171010, -0.173648), vec3(0.969846, 0.171010, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.925417, 0.336824, -0.173648), vec3(0.925417, 0.336824, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.925417, 0.163176, -0.342020), vec3(0.925417, 0.163176, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.883022, 0.321394, -0.342020), vec3(0.883022, 0.321394, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.852869, 0.150384, -0.500000), vec3(0.852869, 0.150384, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.813798, 0.296198, -0.500000), vec3(0.813798, 0.296198, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.754407, 0.133022, -0.642788), vec3(0.754407, 0.133022, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.719846, 0.262003, -0.642788), vec3(0.719846, 0.262003, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.633022, 0.111619, -0.766044), vec3(0.633022, 0.111619, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.604023, 0.219846, -0.766044), vec3(0.604023, 0.219846, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, 0.086824, -0.866025), vec3(0.492404, 0.086824, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.469846, 0.171010, -0.866025), vec3(0.469846, 0.171010, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.336824, 0.059391, -0.939693), vec3(0.336824, 0.059391, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, 0.116978, -0.939693), vec3(0.321394, 0.116978, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, 0.030154, -0.984808), vec3(0.171010, 0.030154, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.163176, 0.059391, -0.984808), vec3(0.163176, 0.059391, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, -1.000000), vec3(0.000000, 0.000000, -1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.163176, 0.059391, -0.984808), vec3(0.163176, 0.059391, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.150384, 0.086824, -0.984808), vec3(0.150384, 0.086824, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, 0.116978, -0.939693), vec3(0.321394, 0.116978, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.296198, 0.171010, -0.939693), vec3(0.296198, 0.171010, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.469846, 0.171010, -0.866025), vec3(0.469846, 0.171010, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.433013, 0.250000, -0.866025), vec3(0.433013, 0.250000, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.604023, 0.219846, -0.766044), vec3(0.604023, 0.219846, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.556670, 0.321394, -0.766044), vec3(0.556670, 0.321394, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.719846, 0.262003, -0.642788), vec3(0.719846, 0.262003, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.663414, 0.383022, -0.642788), vec3(0.663414, 0.383022, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.813798, 0.296198, -0.500000), vec3(0.813798, 0.296198, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.750000, 0.433013, -0.500000), vec3(0.750000, 0.433013, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.883022, 0.321394, -0.342020), vec3(0.883022, 0.321394, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.813798, 0.469846, -0.342020), vec3(0.813798, 0.469846, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.925417, 0.336824, -0.173648), vec3(0.925417, 0.336824, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.852869, 0.492404, -0.173648), vec3(0.852869, 0.492404, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.939693, 0.342020, 0.000000), vec3(0.939693, 0.342020, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.866025, 0.500000, 0.000000), vec3(0.866025, 0.500000, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.925417, 0.336824, 0.173648), vec3(0.925417, 0.336824, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.852869, 0.492404, 0.173648), vec3(0.852869, 0.492404, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.883022, 0.321394, 0.342020), vec3(0.883022, 0.321394, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.813798, 0.469846, 0.342020), vec3(0.813798, 0.469846, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.813798, 0.296198, 0.500000), vec3(0.813798, 0.296198, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.750000, 0.433013, 0.500000), vec3(0.750000, 0.433013, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.719846, 0.262003, 0.642788), vec3(0.719846, 0.262003, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.663414, 0.383022, 0.642788), vec3(0.663414, 0.383022, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.604023, 0.219846, 0.766044), vec3(0.604023, 0.219846, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.556670, 0.321394, 0.766044), vec3(0.556670, 0.321394, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.469846, 0.171010, 0.866025), vec3(0.469846, 0.171010, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.433013, 0.250000, 0.866025), vec3(0.433013, 0.250000, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, 0.116978, 0.939693), vec3(0.321394, 0.116978, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.296198, 0.171010, 0.939693), vec3(0.296198, 0.171010, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.163176, 0.059391, 0.984808), vec3(0.163176, 0.059391, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.150384, 0.086824, 0.984808), vec3(0.150384, 0.086824, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, 1.000000), vec3(0.000000, 0.000000, 1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.150384, 0.086824, 0.984808), vec3(0.150384, 0.086824, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.133022, 0.111619, 0.984808), vec3(0.133022, 0.111619, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.296198, 0.171010, 0.939693), vec3(0.296198, 0.171010, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.262003, 0.219846, 0.939693), vec3(0.262003, 0.219846, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.433013, 0.250000, 0.866025), vec3(0.433013, 0.250000, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.383022, 0.321394, 0.866025), vec3(0.383022, 0.321394, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.556670, 0.321394, 0.766044), vec3(0.556670, 0.321394, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, 0.413176, 0.766044), vec3(0.492404, 0.413176, 0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.663414, 0.383022, 0.642788), vec3(0.663414, 0.383022, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.586824, 0.492404, 0.642788), vec3(0.586824, 0.492404, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.750000, 0.433013, 0.500000), vec3(0.750000, 0.433013, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.663414, 0.556670, 0.500000), vec3(0.663414, 0.556670, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.813798, 0.469846, 0.342020), vec3(0.813798, 0.469846, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.719846, 0.604023, 0.342020), vec3(0.719846, 0.604023, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.852869, 0.492404, 0.173648), vec3(0.852869, 0.492404, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.754407, 0.633022, 0.173648), vec3(0.754407, 0.633022, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.866025, 0.500000, 0.000000), vec3(0.866025, 0.500000, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.766044, 0.642788, 0.000000), vec3(0.766044, 0.642788, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.852869, 0.492404, -0.173648), vec3(0.852869, 0.492404, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.754407, 0.633022, -0.173648), vec3(0.754407, 0.633022, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.813798, 0.469846, -0.342020), vec3(0.813798, 0.469846, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.719846, 0.604023, -0.342020), vec3(0.719846, 0.604023, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.750000, 0.433013, -0.500000), vec3(0.750000, 0.433013, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.663414, 0.556670, -0.500000), vec3(0.663414, 0.556670, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.663414, 0.383022, -0.642788), vec3(0.663414, 0.383022, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.586824, 0.492404, -0.642788), vec3(0.586824, 0.492404, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.556670, 0.321394, -0.766044), vec3(0.556670, 0.321394, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, 0.413176, -0.766044), vec3(0.492404, 0.413176, -0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.433013, 0.250000, -0.866025), vec3(0.433013, 0.250000, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.383022, 0.321394, -0.866025), vec3(0.383022, 0.321394, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.296198, 0.171010, -0.939693), vec3(0.296198, 0.171010, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.262003, 0.219846, -0.939693), vec3(0.262003, 0.219846, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.150384, 0.086824, -0.984808), vec3(0.150384, 0.086824, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.133022, 0.111619, -0.984808), vec3(0.133022, 0.111619, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, -1.000000), vec3(0.000000, 0.000000, -1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.133022, 0.111619, -0.984808), vec3(0.133022, 0.111619, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.111619, 0.133022, -0.984808), vec3(0.111619, 0.133022, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.262003, 0.219846, -0.939693), vec3(0.262003, 0.219846, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.219846, 0.262003, -0.939693), vec3(0.219846, 0.262003, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.383022, 0.321394, -0.866025), vec3(0.383022, 0.321394, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, 0.383022, -0.866025), vec3(0.321394, 0.383022, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, 0.413176, -0.766044), vec3(0.492404, 0.413176, -0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.413176, 0.492404, -0.766044), vec3(0.413176, 0.492404, -0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.586824, 0.492404, -0.642788), vec3(0.586824, 0.492404, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, 0.586824, -0.642788), vec3(0.492404, 0.586824, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.663414, 0.556670, -0.500000), vec3(0.663414, 0.556670, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.556670, 0.663414, -0.500000), vec3(0.556670, 0.663414, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.719846, 0.604023, -0.342020), vec3(0.719846, 0.604023, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.604023, 0.719846, -0.342020), vec3(0.604023, 0.719846, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.754407, 0.633022, -0.173648), vec3(0.754407, 0.633022, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.633022, 0.754407, -0.173648), vec3(0.633022, 0.754407, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.766044, 0.642788, 0.000000), vec3(0.766044, 0.642788, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.642788, 0.766044, 0.000000), vec3(0.642788, 0.766044, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.754407, 0.633022, 0.173648), vec3(0.754407, 0.633022, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.633022, 0.754407, 0.173648), vec3(0.633022, 0.754407, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.719846, 0.604023, 0.342020), vec3(0.719846, 0.604023, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.604023, 0.719846, 0.342020), vec3(0.604023, 0.719846, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.663414, 0.556670, 0.500000), vec3(0.663414, 0.556670, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.556670, 0.663414, 0.500000), vec3(0.556670, 0.663414, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.586824, 0.492404, 0.642788), vec3(0.586824, 0.492404, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, 0.586824, 0.642788), vec3(0.492404, 0.586824, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, 0.413176, 0.766044), vec3(0.492404, 0.413176, 0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.413176, 0.492404, 0.766044), vec3(0.413176, 0.492404, 0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.383022, 0.321394, 0.866025), vec3(0.383022, 0.321394, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, 0.383022, 0.866025), vec3(0.321394, 0.383022, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.262003, 0.219846, 0.939693), vec3(0.262003, 0.219846, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.219846, 0.262003, 0.939693), vec3(0.219846, 0.262003, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.133022, 0.111619, 0.984808), vec3(0.133022, 0.111619, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.111619, 0.133022, 0.984808), vec3(0.111619, 0.133022, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, 1.000000), vec3(0.000000, 0.000000, 1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.111619, 0.133022, 0.984808), vec3(0.111619, 0.133022, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.086824, 0.150384, 0.984808), vec3(0.086824, 0.150384, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.219846, 0.262003, 0.939693), vec3(0.219846, 0.262003, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, 0.296198, 0.939693), vec3(0.171010, 0.296198, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, 0.383022, 0.866025), vec3(0.321394, 0.383022, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.250000, 0.433013, 0.866025), vec3(0.250000, 0.433013, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.413176, 0.492404, 0.766044), vec3(0.413176, 0.492404, 0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, 0.556670, 0.766044), vec3(0.321394, 0.556670, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, 0.586824, 0.642788), vec3(0.492404, 0.586824, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.383022, 0.663414, 0.642788), vec3(0.383022, 0.663414, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.556670, 0.663414, 0.500000), vec3(0.556670, 0.663414, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.433013, 0.750000, 0.500000), vec3(0.433013, 0.750000, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.604023, 0.719846, 0.342020), vec3(0.604023, 0.719846, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.469846, 0.813798, 0.342020), vec3(0.469846, 0.813798, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.633022, 0.754407, 0.173648), vec3(0.633022, 0.754407, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, 0.852869, 0.173648), vec3(0.492404, 0.852869, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.642788, 0.766044, 0.000000), vec3(0.642788, 0.766044, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.500000, 0.866025, 0.000000), vec3(0.500000, 0.866025, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.633022, 0.754407, -0.173648), vec3(0.633022, 0.754407, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, 0.852869, -0.173648), vec3(0.492404, 0.852869, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.604023, 0.719846, -0.342020), vec3(0.604023, 0.719846, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.469846, 0.813798, -0.342020), vec3(0.469846, 0.813798, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.556670, 0.663414, -0.500000), vec3(0.556670, 0.663414, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.433013, 0.750000, -0.500000), vec3(0.433013, 0.750000, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, 0.586824, -0.642788), vec3(0.492404, 0.586824, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.383022, 0.663414, -0.642788), vec3(0.383022, 0.663414, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.413176, 0.492404, -0.766044), vec3(0.413176, 0.492404, -0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, 0.556670, -0.766044), vec3(0.321394, 0.556670, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, 0.383022, -0.866025), vec3(0.321394, 0.383022, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.250000, 0.433013, -0.866025), vec3(0.250000, 0.433013, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.219846, 0.262003, -0.939693), vec3(0.219846, 0.262003, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, 0.296198, -0.939693), vec3(0.171010, 0.296198, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.111619, 0.133022, -0.984808), vec3(0.111619, 0.133022, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.086824, 0.150384, -0.984808), vec3(0.086824, 0.150384, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, -1.000000), vec3(0.000000, 0.000000, -1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.086824, 0.150384, -0.984808), vec3(0.086824, 0.150384, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.059391, 0.163176, -0.984808), vec3(0.059391, 0.163176, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, 0.296198, -0.939693), vec3(0.171010, 0.296198, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.116978, 0.321394, -0.939693), vec3(0.116978, 0.321394, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.250000, 0.433013, -0.866025), vec3(0.250000, 0.433013, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, 0.469846, -0.866025), vec3(0.171010, 0.469846, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, 0.556670, -0.766044), vec3(0.321394, 0.556670, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.219846, 0.604023, -0.766044), vec3(0.219846, 0.604023, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.383022, 0.663414, -0.642788), vec3(0.383022, 0.663414, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.262003, 0.719846, -0.642788), vec3(0.262003, 0.719846, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.433013, 0.750000, -0.500000), vec3(0.433013, 0.750000, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.296198, 0.813798, -0.500000), vec3(0.296198, 0.813798, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.469846, 0.813798, -0.342020), vec3(0.469846, 0.813798, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, 0.883022, -0.342020), vec3(0.321394, 0.883022, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, 0.852869, -0.173648), vec3(0.492404, 0.852869, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.336824, 0.925417, -0.173648), vec3(0.336824, 0.925417, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.500000, 0.866025, 0.000000), vec3(0.500000, 0.866025, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.342020, 0.939693, 0.000000), vec3(0.342020, 0.939693, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, 0.852869, 0.173648), vec3(0.492404, 0.852869, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.336824, 0.925417, 0.173648), vec3(0.336824, 0.925417, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.469846, 0.813798, 0.342020), vec3(0.469846, 0.813798, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, 0.883022, 0.342020), vec3(0.321394, 0.883022, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.433013, 0.750000, 0.500000), vec3(0.433013, 0.750000, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.296198, 0.813798, 0.500000), vec3(0.296198, 0.813798, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.383022, 0.663414, 0.642788), vec3(0.383022, 0.663414, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.262003, 0.719846, 0.642788), vec3(0.262003, 0.719846, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, 0.556670, 0.766044), vec3(0.321394, 0.556670, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.219846, 0.604023, 0.766044), vec3(0.219846, 0.604023, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.250000, 0.433013, 0.866025), vec3(0.250000, 0.433013, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, 0.469846, 0.866025), vec3(0.171010, 0.469846, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, 0.296198, 0.939693), vec3(0.171010, 0.296198, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.116978, 0.321394, 0.939693), vec3(0.116978, 0.321394, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.086824, 0.150384, 0.984808), vec3(0.086824, 0.150384, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.059391, 0.163176, 0.984808), vec3(0.059391, 0.163176, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, 1.000000), vec3(0.000000, 0.000000, 1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.059391, 0.163176, 0.984808), vec3(0.059391, 0.163176, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.030154, 0.171010, 0.984808), vec3(0.030154, 0.171010, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.116978, 0.321394, 0.939693), vec3(0.116978, 0.321394, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.059391, 0.336824, 0.939693), vec3(0.059391, 0.336824, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, 0.469846, 0.866025), vec3(0.171010, 0.469846, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.086824, 0.492404, 0.866025), vec3(0.086824, 0.492404, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.219846, 0.604023, 0.766044), vec3(0.219846, 0.604023, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.111619, 0.633022, 0.766044), vec3(0.111619, 0.633022, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.262003, 0.719846, 0.642788), vec3(0.262003, 0.719846, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.133022, 0.754407, 0.642788), vec3(0.133022, 0.754407, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.296198, 0.813798, 0.500000), vec3(0.296198, 0.813798, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.150384, 0.852869, 0.500000), vec3(0.150384, 0.852869, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, 0.883022, 0.342020), vec3(0.321394, 0.883022, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.163176, 0.925417, 0.342020), vec3(0.163176, 0.925417, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.336824, 0.925417, 0.173648), vec3(0.336824, 0.925417, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, 0.969846, 0.173648), vec3(0.171010, 0.969846, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.342020, 0.939693, 0.000000), vec3(0.342020, 0.939693, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.173648, 0.984808, 0.000000), vec3(0.173648, 0.984808, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.336824, 0.925417, -0.173648), vec3(0.336824, 0.925417, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, 0.969846, -0.173648), vec3(0.171010, 0.969846, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, 0.883022, -0.342020), vec3(0.321394, 0.883022, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.163176, 0.925417, -0.342020), vec3(0.163176, 0.925417, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.296198, 0.813798, -0.500000), vec3(0.296198, 0.813798, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.150384, 0.852869, -0.500000), vec3(0.150384, 0.852869, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.262003, 0.719846, -0.642788), vec3(0.262003, 0.719846, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.133022, 0.754407, -0.642788), vec3(0.133022, 0.754407, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.219846, 0.604023, -0.766044), vec3(0.219846, 0.604023, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.111619, 0.633022, -0.766044), vec3(0.111619, 0.633022, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, 0.469846, -0.866025), vec3(0.171010, 0.469846, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.086824, 0.492404, -0.866025), vec3(0.086824, 0.492404, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.116978, 0.321394, -0.939693), vec3(0.116978, 0.321394, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.059391, 0.336824, -0.939693), vec3(0.059391, 0.336824, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.059391, 0.163176, -0.984808), vec3(0.059391, 0.163176, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.030154, 0.171010, -0.984808), vec3(0.030154, 0.171010, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, -1.000000), vec3(0.000000, 0.000000, -1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.030154, 0.171010, -0.984808), vec3(0.030154, 0.171010, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.173648, -0.984808), vec3(0.000000, 0.173648, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.059391, 0.336824, -0.939693), vec3(0.059391, 0.336824, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.342020, -0.939693), vec3(0.000000, 0.342020, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.086824, 0.492404, -0.866025), vec3(0.086824, 0.492404, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.500000, -0.866025), vec3(0.000000, 0.500000, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.111619, 0.633022, -0.766044), vec3(0.111619, 0.633022, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.642788, -0.766044), vec3(0.000000, 0.642788, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.133022, 0.754407, -0.642788), vec3(0.133022, 0.754407, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.766044, -0.642788), vec3(0.000000, 0.766044, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.150384, 0.852869, -0.500000), vec3(0.150384, 0.852869, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.866025, -0.500000), vec3(0.000000, 0.866025, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.163176, 0.925417, -0.342020), vec3(0.163176, 0.925417, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.939693, -0.342020), vec3(0.000000, 0.939693, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, 0.969846, -0.173648), vec3(0.171010, 0.969846, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.984808, -0.173648), vec3(0.000000, 0.984808, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.173648, 0.984808, 0.000000), vec3(0.173648, 0.984808, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 1.000000, 0.000000), vec3(0.000000, 1.000000, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, 0.969846, 0.173648), vec3(0.171010, 0.969846, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.984808, 0.173648), vec3(0.000000, 0.984808, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.163176, 0.925417, 0.342020), vec3(0.163176, 0.925417, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.939693, 0.342020), vec3(0.000000, 0.939693, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.150384, 0.852869, 0.500000), vec3(0.150384, 0.852869, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.866025, 0.500000), vec3(0.000000, 0.866025, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.133022, 0.754407, 0.642788), vec3(0.133022, 0.754407, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.766044, 0.642788), vec3(0.000000, 0.766044, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.111619, 0.633022, 0.766044), vec3(0.111619, 0.633022, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.642788, 0.766044), vec3(0.000000, 0.642788, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.086824, 0.492404, 0.866025), vec3(0.086824, 0.492404, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.500000, 0.866025), vec3(0.000000, 0.500000, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.059391, 0.336824, 0.939693), vec3(0.059391, 0.336824, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.342020, 0.939693), vec3(0.000000, 0.342020, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.030154, 0.171010, 0.984808), vec3(0.030154, 0.171010, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.173648, 0.984808), vec3(0.000000, 0.173648, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, 1.000000), vec3(0.000000, 0.000000, 1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.173648, 0.984808), vec3(0.000000, 0.173648, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.030154, 0.171010, 0.984808), vec3(-0.030154, 0.171010, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.342020, 0.939693), vec3(0.000000, 0.342020, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.059391, 0.336824, 0.939693), vec3(-0.059391, 0.336824, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.500000, 0.866025), vec3(0.000000, 0.500000, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.086824, 0.492404, 0.866025), vec3(-0.086824, 0.492404, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.642788, 0.766044), vec3(0.000000, 0.642788, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.111619, 0.633022, 0.766044), vec3(-0.111619, 0.633022, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.766044, 0.642788), vec3(0.000000, 0.766044, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.133022, 0.754407, 0.642788), vec3(-0.133022, 0.754407, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.866025, 0.500000), vec3(0.000000, 0.866025, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.150384, 0.852869, 0.500000), vec3(-0.150384, 0.852869, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.939693, 0.342020), vec3(0.000000, 0.939693, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.163176, 0.925417, 0.342020), vec3(-0.163176, 0.925417, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.984808, 0.173648), vec3(0.000000, 0.984808, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, 0.969846, 0.173648), vec3(-0.171010, 0.969846, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 1.000000, 0.000000), vec3(0.000000, 1.000000, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.173648, 0.984808, 0.000000), vec3(-0.173648, 0.984808, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.984808, -0.173648), vec3(0.000000, 0.984808, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, 0.969846, -0.173648), vec3(-0.171010, 0.969846, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.939693, -0.342020), vec3(0.000000, 0.939693, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.163176, 0.925417, -0.342020), vec3(-0.163176, 0.925417, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.866025, -0.500000), vec3(0.000000, 0.866025, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.150384, 0.852869, -0.500000), vec3(-0.150384, 0.852869, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.766044, -0.642788), vec3(0.000000, 0.766044, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.133022, 0.754407, -0.642788), vec3(-0.133022, 0.754407, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.642788, -0.766044), vec3(0.000000, 0.642788, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.111619, 0.633022, -0.766044), vec3(-0.111619, 0.633022, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.500000, -0.866025), vec3(0.000000, 0.500000, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.086824, 0.492404, -0.866025), vec3(-0.086824, 0.492404, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.342020, -0.939693), vec3(0.000000, 0.342020, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.059391, 0.336824, -0.939693), vec3(-0.059391, 0.336824, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.173648, -0.984808), vec3(0.000000, 0.173648, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.030154, 0.171010, -0.984808), vec3(-0.030154, 0.171010, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, -1.000000), vec3(0.000000, 0.000000, -1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.030154, 0.171010, -0.984808), vec3(-0.030154, 0.171010, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.059391, 0.163176, -0.984808), vec3(-0.059391, 0.163176, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.059391, 0.336824, -0.939693), vec3(-0.059391, 0.336824, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.116978, 0.321394, -0.939693), vec3(-0.116978, 0.321394, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.086824, 0.492404, -0.866025), vec3(-0.086824, 0.492404, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, 0.469846, -0.866025), vec3(-0.171010, 0.469846, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.111619, 0.633022, -0.766044), vec3(-0.111619, 0.633022, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.219846, 0.604023, -0.766044), vec3(-0.219846, 0.604023, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.133022, 0.754407, -0.642788), vec3(-0.133022, 0.754407, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.262003, 0.719846, -0.642788), vec3(-0.262003, 0.719846, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.150384, 0.852869, -0.500000), vec3(-0.150384, 0.852869, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.296198, 0.813798, -0.500000), vec3(-0.296198, 0.813798, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.163176, 0.925417, -0.342020), vec3(-0.163176, 0.925417, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, 0.883022, -0.342020), vec3(-0.321394, 0.883022, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, 0.969846, -0.173648), vec3(-0.171010, 0.969846, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.336824, 0.925417, -0.173648), vec3(-0.336824, 0.925417, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.173648, 0.984808, 0.000000), vec3(-0.173648, 0.984808, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.342020, 0.939693, 0.000000), vec3(-0.342020, 0.939693, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, 0.969846, 0.173648), vec3(-0.171010, 0.969846, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.336824, 0.925417, 0.173648), vec3(-0.336824, 0.925417, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.163176, 0.925417, 0.342020), vec3(-0.163176, 0.925417, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, 0.883022, 0.342020), vec3(-0.321394, 0.883022, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.150384, 0.852869, 0.500000), vec3(-0.150384, 0.852869, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.296198, 0.813798, 0.500000), vec3(-0.296198, 0.813798, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.133022, 0.754407, 0.642788), vec3(-0.133022, 0.754407, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.262003, 0.719846, 0.642788), vec3(-0.262003, 0.719846, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.111619, 0.633022, 0.766044), vec3(-0.111619, 0.633022, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.219846, 0.604023, 0.766044), vec3(-0.219846, 0.604023, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.086824, 0.492404, 0.866025), vec3(-0.086824, 0.492404, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, 0.469846, 0.866025), vec3(-0.171010, 0.469846, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.059391, 0.336824, 0.939693), vec3(-0.059391, 0.336824, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.116978, 0.321394, 0.939693), vec3(-0.116978, 0.321394, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.030154, 0.171010, 0.984808), vec3(-0.030154, 0.171010, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.059391, 0.163176, 0.984808), vec3(-0.059391, 0.163176, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, 1.000000), vec3(0.000000, 0.000000, 1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.059391, 0.163176, 0.984808), vec3(-0.059391, 0.163176, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.086824, 0.150384, 0.984808), vec3(-0.086824, 0.150384, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.116978, 0.321394, 0.939693), vec3(-0.116978, 0.321394, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, 0.296198, 0.939693), vec3(-0.171010, 0.296198, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, 0.469846, 0.866025), vec3(-0.171010, 0.469846, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.250000, 0.433013, 0.866025), vec3(-0.250000, 0.433013, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.219846, 0.604023, 0.766044), vec3(-0.219846, 0.604023, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, 0.556670, 0.766044), vec3(-0.321394, 0.556670, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.262003, 0.719846, 0.642788), vec3(-0.262003, 0.719846, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.383022, 0.663414, 0.642788), vec3(-0.383022, 0.663414, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.296198, 0.813798, 0.500000), vec3(-0.296198, 0.813798, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.433013, 0.750000, 0.500000), vec3(-0.433013, 0.750000, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, 0.883022, 0.342020), vec3(-0.321394, 0.883022, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.469846, 0.813798, 0.342020), vec3(-0.469846, 0.813798, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.336824, 0.925417, 0.173648), vec3(-0.336824, 0.925417, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, 0.852869, 0.173648), vec3(-0.492404, 0.852869, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.342020, 0.939693, 0.000000), vec3(-0.342020, 0.939693, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.500000, 0.866025, 0.000000), vec3(-0.500000, 0.866025, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.336824, 0.925417, -0.173648), vec3(-0.336824, 0.925417, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, 0.852869, -0.173648), vec3(-0.492404, 0.852869, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, 0.883022, -0.342020), vec3(-0.321394, 0.883022, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.469846, 0.813798, -0.342020), vec3(-0.469846, 0.813798, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.296198, 0.813798, -0.500000), vec3(-0.296198, 0.813798, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.433013, 0.750000, -0.500000), vec3(-0.433013, 0.750000, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.262003, 0.719846, -0.642788), vec3(-0.262003, 0.719846, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.383022, 0.663414, -0.642788), vec3(-0.383022, 0.663414, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.219846, 0.604023, -0.766044), vec3(-0.219846, 0.604023, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, 0.556670, -0.766044), vec3(-0.321394, 0.556670, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, 0.469846, -0.866025), vec3(-0.171010, 0.469846, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.250000, 0.433013, -0.866025), vec3(-0.250000, 0.433013, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.116978, 0.321394, -0.939693), vec3(-0.116978, 0.321394, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, 0.296198, -0.939693), vec3(-0.171010, 0.296198, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.059391, 0.163176, -0.984808), vec3(-0.059391, 0.163176, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.086824, 0.150384, -0.984808), vec3(-0.086824, 0.150384, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, -1.000000), vec3(0.000000, 0.000000, -1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.086824, 0.150384, -0.984808), vec3(-0.086824, 0.150384, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.111619, 0.133022, -0.984808), vec3(-0.111619, 0.133022, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, 0.296198, -0.939693), vec3(-0.171010, 0.296198, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.219846, 0.262003, -0.939693), vec3(-0.219846, 0.262003, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.250000, 0.433013, -0.866025), vec3(-0.250000, 0.433013, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, 0.383022, -0.866025), vec3(-0.321394, 0.383022, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, 0.556670, -0.766044), vec3(-0.321394, 0.556670, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.413176, 0.492404, -0.766044), vec3(-0.413176, 0.492404, -0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.383022, 0.663414, -0.642788), vec3(-0.383022, 0.663414, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, 0.586824, -0.642788), vec3(-0.492404, 0.586824, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.433013, 0.750000, -0.500000), vec3(-0.433013, 0.750000, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.556670, 0.663414, -0.500000), vec3(-0.556670, 0.663414, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.469846, 0.813798, -0.342020), vec3(-0.469846, 0.813798, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.604023, 0.719846, -0.342020), vec3(-0.604023, 0.719846, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, 0.852869, -0.173648), vec3(-0.492404, 0.852869, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.633022, 0.754407, -0.173648), vec3(-0.633022, 0.754407, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.500000, 0.866025, 0.000000), vec3(-0.500000, 0.866025, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.642788, 0.766044, 0.000000), vec3(-0.642788, 0.766044, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, 0.852869, 0.173648), vec3(-0.492404, 0.852869, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.633022, 0.754407, 0.173648), vec3(-0.633022, 0.754407, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.469846, 0.813798, 0.342020), vec3(-0.469846, 0.813798, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.604023, 0.719846, 0.342020), vec3(-0.604023, 0.719846, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.433013, 0.750000, 0.500000), vec3(-0.433013, 0.750000, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.556670, 0.663414, 0.500000), vec3(-0.556670, 0.663414, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.383022, 0.663414, 0.642788), vec3(-0.383022, 0.663414, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, 0.586824, 0.642788), vec3(-0.492404, 0.586824, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, 0.556670, 0.766044), vec3(-0.321394, 0.556670, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.413176, 0.492404, 0.766044), vec3(-0.413176, 0.492404, 0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.250000, 0.433013, 0.866025), vec3(-0.250000, 0.433013, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, 0.383022, 0.866025), vec3(-0.321394, 0.383022, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, 0.296198, 0.939693), vec3(-0.171010, 0.296198, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.219846, 0.262003, 0.939693), vec3(-0.219846, 0.262003, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.086824, 0.150384, 0.984808), vec3(-0.086824, 0.150384, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.111619, 0.133022, 0.984808), vec3(-0.111619, 0.133022, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, 1.000000), vec3(0.000000, 0.000000, 1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.111619, 0.133022, 0.984808), vec3(-0.111619, 0.133022, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.133022, 0.111619, 0.984808), vec3(-0.133022, 0.111619, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.219846, 0.262003, 0.939693), vec3(-0.219846, 0.262003, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.262003, 0.219846, 0.939693), vec3(-0.262003, 0.219846, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, 0.383022, 0.866025), vec3(-0.321394, 0.383022, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.383022, 0.321394, 0.866025), vec3(-0.383022, 0.321394, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.413176, 0.492404, 0.766044), vec3(-0.413176, 0.492404, 0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, 0.413176, 0.766044), vec3(-0.492404, 0.413176, 0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, 0.586824, 0.642788), vec3(-0.492404, 0.586824, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.586824, 0.492404, 0.642788), vec3(-0.586824, 0.492404, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.556670, 0.663414, 0.500000), vec3(-0.556670, 0.663414, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.663414, 0.556670, 0.500000), vec3(-0.663414, 0.556670, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.604023, 0.719846, 0.342020), vec3(-0.604023, 0.719846, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.719846, 0.604023, 0.342020), vec3(-0.719846, 0.604023, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.633022, 0.754407, 0.173648), vec3(-0.633022, 0.754407, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.754407, 0.633022, 0.173648), vec3(-0.754407, 0.633022, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.642788, 0.766044, 0.000000), vec3(-0.642788, 0.766044, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.766044, 0.642788, 0.000000), vec3(-0.766044, 0.642788, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.633022, 0.754407, -0.173648), vec3(-0.633022, 0.754407, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.754407, 0.633022, -0.173648), vec3(-0.754407, 0.633022, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.604023, 0.719846, -0.342020), vec3(-0.604023, 0.719846, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.719846, 0.604023, -0.342020), vec3(-0.719846, 0.604023, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.556670, 0.663414, -0.500000), vec3(-0.556670, 0.663414, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.663414, 0.556670, -0.500000), vec3(-0.663414, 0.556670, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, 0.586824, -0.642788), vec3(-0.492404, 0.586824, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.586824, 0.492404, -0.642788), vec3(-0.586824, 0.492404, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.413176, 0.492404, -0.766044), vec3(-0.413176, 0.492404, -0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, 0.413176, -0.766044), vec3(-0.492404, 0.413176, -0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, 0.383022, -0.866025), vec3(-0.321394, 0.383022, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.383022, 0.321394, -0.866025), vec3(-0.383022, 0.321394, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.219846, 0.262003, -0.939693), vec3(-0.219846, 0.262003, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.262003, 0.219846, -0.939693), vec3(-0.262003, 0.219846, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.111619, 0.133022, -0.984808), vec3(-0.111619, 0.133022, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.133022, 0.111619, -0.984808), vec3(-0.133022, 0.111619, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, -1.000000), vec3(0.000000, 0.000000, -1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.133022, 0.111619, -0.984808), vec3(-0.133022, 0.111619, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.150384, 0.086824, -0.984808), vec3(-0.150384, 0.086824, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.262003, 0.219846, -0.939693), vec3(-0.262003, 0.219846, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.296198, 0.171010, -0.939693), vec3(-0.296198, 0.171010, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.383022, 0.321394, -0.866025), vec3(-0.383022, 0.321394, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.433013, 0.250000, -0.866025), vec3(-0.433013, 0.250000, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, 0.413176, -0.766044), vec3(-0.492404, 0.413176, -0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.556670, 0.321394, -0.766044), vec3(-0.556670, 0.321394, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.586824, 0.492404, -0.642788), vec3(-0.586824, 0.492404, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.663414, 0.383022, -0.642788), vec3(-0.663414, 0.383022, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.663414, 0.556670, -0.500000), vec3(-0.663414, 0.556670, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.750000, 0.433013, -0.500000), vec3(-0.750000, 0.433013, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.719846, 0.604023, -0.342020), vec3(-0.719846, 0.604023, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.813798, 0.469846, -0.342020), vec3(-0.813798, 0.469846, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.754407, 0.633022, -0.173648), vec3(-0.754407, 0.633022, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.852869, 0.492404, -0.173648), vec3(-0.852869, 0.492404, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.766044, 0.642788, 0.000000), vec3(-0.766044, 0.642788, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.866025, 0.500000, 0.000000), vec3(-0.866025, 0.500000, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.754407, 0.633022, 0.173648), vec3(-0.754407, 0.633022, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.852869, 0.492404, 0.173648), vec3(-0.852869, 0.492404, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.719846, 0.604023, 0.342020), vec3(-0.719846, 0.604023, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.813798, 0.469846, 0.342020), vec3(-0.813798, 0.469846, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.663414, 0.556670, 0.500000), vec3(-0.663414, 0.556670, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.750000, 0.433013, 0.500000), vec3(-0.750000, 0.433013, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.586824, 0.492404, 0.642788), vec3(-0.586824, 0.492404, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.663414, 0.383022, 0.642788), vec3(-0.663414, 0.383022, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, 0.413176, 0.766044), vec3(-0.492404, 0.413176, 0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.556670, 0.321394, 0.766044), vec3(-0.556670, 0.321394, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.383022, 0.321394, 0.866025), vec3(-0.383022, 0.321394, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.433013, 0.250000, 0.866025), vec3(-0.433013, 0.250000, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.262003, 0.219846, 0.939693), vec3(-0.262003, 0.219846, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.296198, 0.171010, 0.939693), vec3(-0.296198, 0.171010, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.133022, 0.111619, 0.984808), vec3(-0.133022, 0.111619, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.150384, 0.086824, 0.984808), vec3(-0.150384, 0.086824, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, 1.000000), vec3(0.000000, 0.000000, 1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.150384, 0.086824, 0.984808), vec3(-0.150384, 0.086824, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.163176, 0.059391, 0.984808), vec3(-0.163176, 0.059391, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.296198, 0.171010, 0.939693), vec3(-0.296198, 0.171010, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, 0.116978, 0.939693), vec3(-0.321394, 0.116978, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.433013, 0.250000, 0.866025), vec3(-0.433013, 0.250000, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.469846, 0.171010, 0.866025), vec3(-0.469846, 0.171010, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.556670, 0.321394, 0.766044), vec3(-0.556670, 0.321394, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.604023, 0.219846, 0.766044), vec3(-0.604023, 0.219846, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.663414, 0.383022, 0.642788), vec3(-0.663414, 0.383022, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.719846, 0.262003, 0.642788), vec3(-0.719846, 0.262003, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.750000, 0.433013, 0.500000), vec3(-0.750000, 0.433013, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.813798, 0.296198, 0.500000), vec3(-0.813798, 0.296198, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.813798, 0.469846, 0.342020), vec3(-0.813798, 0.469846, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.883022, 0.321394, 0.342020), vec3(-0.883022, 0.321394, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.852869, 0.492404, 0.173648), vec3(-0.852869, 0.492404, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.925417, 0.336824, 0.173648), vec3(-0.925417, 0.336824, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.866025, 0.500000, 0.000000), vec3(-0.866025, 0.500000, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.939693, 0.342020, 0.000000), vec3(-0.939693, 0.342020, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.852869, 0.492404, -0.173648), vec3(-0.852869, 0.492404, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.925417, 0.336824, -0.173648), vec3(-0.925417, 0.336824, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.813798, 0.469846, -0.342020), vec3(-0.813798, 0.469846, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.883022, 0.321394, -0.342020), vec3(-0.883022, 0.321394, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.750000, 0.433013, -0.500000), vec3(-0.750000, 0.433013, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.813798, 0.296198, -0.500000), vec3(-0.813798, 0.296198, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.663414, 0.383022, -0.642788), vec3(-0.663414, 0.383022, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.719846, 0.262003, -0.642788), vec3(-0.719846, 0.262003, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.556670, 0.321394, -0.766044), vec3(-0.556670, 0.321394, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.604023, 0.219846, -0.766044), vec3(-0.604023, 0.219846, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.433013, 0.250000, -0.866025), vec3(-0.433013, 0.250000, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.469846, 0.171010, -0.866025), vec3(-0.469846, 0.171010, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.296198, 0.171010, -0.939693), vec3(-0.296198, 0.171010, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, 0.116978, -0.939693), vec3(-0.321394, 0.116978, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.150384, 0.086824, -0.984808), vec3(-0.150384, 0.086824, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.163176, 0.059391, -0.984808), vec3(-0.163176, 0.059391, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, -1.000000), vec3(0.000000, 0.000000, -1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.163176, 0.059391, -0.984808), vec3(-0.163176, 0.059391, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, 0.030154, -0.984808), vec3(-0.171010, 0.030154, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, 0.116978, -0.939693), vec3(-0.321394, 0.116978, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.336824, 0.059391, -0.939693), vec3(-0.336824, 0.059391, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.469846, 0.171010, -0.866025), vec3(-0.469846, 0.171010, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, 0.086824, -0.866025), vec3(-0.492404, 0.086824, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.604023, 0.219846, -0.766044), vec3(-0.604023, 0.219846, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.633022, 0.111619, -0.766044), vec3(-0.633022, 0.111619, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.719846, 0.262003, -0.642788), vec3(-0.719846, 0.262003, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.754407, 0.133022, -0.642788), vec3(-0.754407, 0.133022, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.813798, 0.296198, -0.500000), vec3(-0.813798, 0.296198, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.852869, 0.150384, -0.500000), vec3(-0.852869, 0.150384, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.883022, 0.321394, -0.342020), vec3(-0.883022, 0.321394, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.925417, 0.163176, -0.342020), vec3(-0.925417, 0.163176, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.925417, 0.336824, -0.173648), vec3(-0.925417, 0.336824, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.969846, 0.171010, -0.173648), vec3(-0.969846, 0.171010, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.939693, 0.342020, 0.000000), vec3(-0.939693, 0.342020, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.984808, 0.173648, 0.000000), vec3(-0.984808, 0.173648, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.925417, 0.336824, 0.173648), vec3(-0.925417, 0.336824, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.969846, 0.171010, 0.173648), vec3(-0.969846, 0.171010, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.883022, 0.321394, 0.342020), vec3(-0.883022, 0.321394, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.925417, 0.163176, 0.342020), vec3(-0.925417, 0.163176, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.813798, 0.296198, 0.500000), vec3(-0.813798, 0.296198, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.852869, 0.150384, 0.500000), vec3(-0.852869, 0.150384, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.719846, 0.262003, 0.642788), vec3(-0.719846, 0.262003, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.754407, 0.133022, 0.642788), vec3(-0.754407, 0.133022, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.604023, 0.219846, 0.766044), vec3(-0.604023, 0.219846, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.633022, 0.111619, 0.766044), vec3(-0.633022, 0.111619, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.469846, 0.171010, 0.866025), vec3(-0.469846, 0.171010, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, 0.086824, 0.866025), vec3(-0.492404, 0.086824, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, 0.116978, 0.939693), vec3(-0.321394, 0.116978, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.336824, 0.059391, 0.939693), vec3(-0.336824, 0.059391, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.163176, 0.059391, 0.984808), vec3(-0.163176, 0.059391, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, 0.030154, 0.984808), vec3(-0.171010, 0.030154, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, 1.000000), vec3(0.000000, 0.000000, 1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, 0.030154, 0.984808), vec3(-0.171010, 0.030154, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.173648, 0.000000, 0.984808), vec3(-0.173648, 0.000000, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.336824, 0.059391, 0.939693), vec3(-0.336824, 0.059391, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.342020, 0.000000, 0.939693), vec3(-0.342020, 0.000000, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, 0.086824, 0.866025), vec3(-0.492404, 0.086824, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.500000, 0.000000, 0.866025), vec3(-0.500000, 0.000000, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.633022, 0.111619, 0.766044), vec3(-0.633022, 0.111619, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.642788, 0.000000, 0.766044), vec3(-0.642788, 0.000000, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.754407, 0.133022, 0.642788), vec3(-0.754407, 0.133022, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.766044, 0.000000, 0.642788), vec3(-0.766044, 0.000000, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.852869, 0.150384, 0.500000), vec3(-0.852869, 0.150384, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.866025, 0.000000, 0.500000), vec3(-0.866025, 0.000000, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.925417, 0.163176, 0.342020), vec3(-0.925417, 0.163176, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.939693, 0.000000, 0.342020), vec3(-0.939693, 0.000000, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.969846, 0.171010, 0.173648), vec3(-0.969846, 0.171010, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.984808, 0.000000, 0.173648), vec3(-0.984808, 0.000000, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.984808, 0.173648, 0.000000), vec3(-0.984808, 0.173648, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-1.000000, 0.000000, 0.000000), vec3(-1.000000, 0.000000, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.969846, 0.171010, -0.173648), vec3(-0.969846, 0.171010, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.984808, 0.000000, -0.173648), vec3(-0.984808, 0.000000, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.925417, 0.163176, -0.342020), vec3(-0.925417, 0.163176, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.939693, 0.000000, -0.342020), vec3(-0.939693, 0.000000, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.852869, 0.150384, -0.500000), vec3(-0.852869, 0.150384, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.866025, 0.000000, -0.500000), vec3(-0.866025, 0.000000, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.754407, 0.133022, -0.642788), vec3(-0.754407, 0.133022, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.766044, 0.000000, -0.642788), vec3(-0.766044, 0.000000, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.633022, 0.111619, -0.766044), vec3(-0.633022, 0.111619, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.642788, 0.000000, -0.766044), vec3(-0.642788, 0.000000, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, 0.086824, -0.866025), vec3(-0.492404, 0.086824, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.500000, 0.000000, -0.866025), vec3(-0.500000, 0.000000, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.336824, 0.059391, -0.939693), vec3(-0.336824, 0.059391, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.342020, 0.000000, -0.939693), vec3(-0.342020, 0.000000, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, 0.030154, -0.984808), vec3(-0.171010, 0.030154, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.173648, 0.000000, -0.984808), vec3(-0.173648, 0.000000, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, -1.000000), vec3(0.000000, 0.000000, -1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.173648, 0.000000, -0.984808), vec3(-0.173648, 0.000000, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, -0.030154, -0.984808), vec3(-0.171010, -0.030154, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.342020, 0.000000, -0.939693), vec3(-0.342020, 0.000000, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.336824, -0.059391, -0.939693), vec3(-0.336824, -0.059391, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.500000, 0.000000, -0.866025), vec3(-0.500000, 0.000000, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, -0.086824, -0.866025), vec3(-0.492404, -0.086824, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.642788, 0.000000, -0.766044), vec3(-0.642788, 0.000000, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.633022, -0.111619, -0.766044), vec3(-0.633022, -0.111619, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.766044, 0.000000, -0.642788), vec3(-0.766044, 0.000000, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.754407, -0.133022, -0.642788), vec3(-0.754407, -0.133022, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.866025, 0.000000, -0.500000), vec3(-0.866025, 0.000000, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.852869, -0.150384, -0.500000), vec3(-0.852869, -0.150384, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.939693, 0.000000, -0.342020), vec3(-0.939693, 0.000000, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.925417, -0.163176, -0.342020), vec3(-0.925417, -0.163176, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.984808, 0.000000, -0.173648), vec3(-0.984808, 0.000000, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.969846, -0.171010, -0.173648), vec3(-0.969846, -0.171010, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-1.000000, 0.000000, 0.000000), vec3(-1.000000, 0.000000, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.984808, -0.173648, 0.000000), vec3(-0.984808, -0.173648, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.984808, 0.000000, 0.173648), vec3(-0.984808, 0.000000, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.969846, -0.171010, 0.173648), vec3(-0.969846, -0.171010, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.939693, 0.000000, 0.342020), vec3(-0.939693, 0.000000, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.925417, -0.163176, 0.342020), vec3(-0.925417, -0.163176, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.866025, 0.000000, 0.500000), vec3(-0.866025, 0.000000, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.852869, -0.150384, 0.500000), vec3(-0.852869, -0.150384, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.766044, 0.000000, 0.642788), vec3(-0.766044, 0.000000, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.754407, -0.133022, 0.642788), vec3(-0.754407, -0.133022, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.642788, 0.000000, 0.766044), vec3(-0.642788, 0.000000, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.633022, -0.111619, 0.766044), vec3(-0.633022, -0.111619, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.500000, 0.000000, 0.866025), vec3(-0.500000, 0.000000, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, -0.086824, 0.866025), vec3(-0.492404, -0.086824, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.342020, 0.000000, 0.939693), vec3(-0.342020, 0.000000, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.336824, -0.059391, 0.939693), vec3(-0.336824, -0.059391, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.173648, 0.000000, 0.984808), vec3(-0.173648, 0.000000, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, -0.030154, 0.984808), vec3(-0.171010, -0.030154, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, 1.000000), vec3(0.000000, 0.000000, 1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, -0.030154, 0.984808), vec3(-0.171010, -0.030154, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.163176, -0.059391, 0.984808), vec3(-0.163176, -0.059391, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.336824, -0.059391, 0.939693), vec3(-0.336824, -0.059391, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, -0.116978, 0.939693), vec3(-0.321394, -0.116978, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, -0.086824, 0.866025), vec3(-0.492404, -0.086824, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.469846, -0.171010, 0.866025), vec3(-0.469846, -0.171010, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.633022, -0.111619, 0.766044), vec3(-0.633022, -0.111619, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.604023, -0.219846, 0.766044), vec3(-0.604023, -0.219846, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.754407, -0.133022, 0.642788), vec3(-0.754407, -0.133022, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.719846, -0.262003, 0.642788), vec3(-0.719846, -0.262003, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.852869, -0.150384, 0.500000), vec3(-0.852869, -0.150384, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.813798, -0.296198, 0.500000), vec3(-0.813798, -0.296198, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.925417, -0.163176, 0.342020), vec3(-0.925417, -0.163176, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.883022, -0.321394, 0.342020), vec3(-0.883022, -0.321394, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.969846, -0.171010, 0.173648), vec3(-0.969846, -0.171010, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.925417, -0.336824, 0.173648), vec3(-0.925417, -0.336824, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.984808, -0.173648, 0.000000), vec3(-0.984808, -0.173648, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.939693, -0.342020, 0.000000), vec3(-0.939693, -0.342020, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.969846, -0.171010, -0.173648), vec3(-0.969846, -0.171010, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.925417, -0.336824, -0.173648), vec3(-0.925417, -0.336824, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.925417, -0.163176, -0.342020), vec3(-0.925417, -0.163176, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.883022, -0.321394, -0.342020), vec3(-0.883022, -0.321394, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.852869, -0.150384, -0.500000), vec3(-0.852869, -0.150384, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.813798, -0.296198, -0.500000), vec3(-0.813798, -0.296198, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.754407, -0.133022, -0.642788), vec3(-0.754407, -0.133022, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.719846, -0.262003, -0.642788), vec3(-0.719846, -0.262003, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.633022, -0.111619, -0.766044), vec3(-0.633022, -0.111619, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.604023, -0.219846, -0.766044), vec3(-0.604023, -0.219846, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, -0.086824, -0.866025), vec3(-0.492404, -0.086824, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.469846, -0.171010, -0.866025), vec3(-0.469846, -0.171010, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.336824, -0.059391, -0.939693), vec3(-0.336824, -0.059391, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, -0.116978, -0.939693), vec3(-0.321394, -0.116978, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, -0.030154, -0.984808), vec3(-0.171010, -0.030154, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.163176, -0.059391, -0.984808), vec3(-0.163176, -0.059391, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, -1.000000), vec3(0.000000, 0.000000, -1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.163176, -0.059391, -0.984808), vec3(-0.163176, -0.059391, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.150384, -0.086824, -0.984808), vec3(-0.150384, -0.086824, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, -0.116978, -0.939693), vec3(-0.321394, -0.116978, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.296198, -0.171010, -0.939693), vec3(-0.296198, -0.171010, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.469846, -0.171010, -0.866025), vec3(-0.469846, -0.171010, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.433013, -0.250000, -0.866025), vec3(-0.433013, -0.250000, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.604023, -0.219846, -0.766044), vec3(-0.604023, -0.219846, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.556670, -0.321394, -0.766044), vec3(-0.556670, -0.321394, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.719846, -0.262003, -0.642788), vec3(-0.719846, -0.262003, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.663414, -0.383022, -0.642788), vec3(-0.663414, -0.383022, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.813798, -0.296198, -0.500000), vec3(-0.813798, -0.296198, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.750000, -0.433013, -0.500000), vec3(-0.750000, -0.433013, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.883022, -0.321394, -0.342020), vec3(-0.883022, -0.321394, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.813798, -0.469846, -0.342020), vec3(-0.813798, -0.469846, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.925417, -0.336824, -0.173648), vec3(-0.925417, -0.336824, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.852869, -0.492404, -0.173648), vec3(-0.852869, -0.492404, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.939693, -0.342020, 0.000000), vec3(-0.939693, -0.342020, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.866025, -0.500000, 0.000000), vec3(-0.866025, -0.500000, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.925417, -0.336824, 0.173648), vec3(-0.925417, -0.336824, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.852869, -0.492404, 0.173648), vec3(-0.852869, -0.492404, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.883022, -0.321394, 0.342020), vec3(-0.883022, -0.321394, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.813798, -0.469846, 0.342020), vec3(-0.813798, -0.469846, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.813798, -0.296198, 0.500000), vec3(-0.813798, -0.296198, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.750000, -0.433013, 0.500000), vec3(-0.750000, -0.433013, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.719846, -0.262003, 0.642788), vec3(-0.719846, -0.262003, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.663414, -0.383022, 0.642788), vec3(-0.663414, -0.383022, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.604023, -0.219846, 0.766044), vec3(-0.604023, -0.219846, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.556670, -0.321394, 0.766044), vec3(-0.556670, -0.321394, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.469846, -0.171010, 0.866025), vec3(-0.469846, -0.171010, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.433013, -0.250000, 0.866025), vec3(-0.433013, -0.250000, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, -0.116978, 0.939693), vec3(-0.321394, -0.116978, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.296198, -0.171010, 0.939693), vec3(-0.296198, -0.171010, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.163176, -0.059391, 0.984808), vec3(-0.163176, -0.059391, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.150384, -0.086824, 0.984808), vec3(-0.150384, -0.086824, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, 1.000000), vec3(0.000000, 0.000000, 1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.150384, -0.086824, 0.984808), vec3(-0.150384, -0.086824, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.133022, -0.111619, 0.984808), vec3(-0.133022, -0.111619, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.296198, -0.171010, 0.939693), vec3(-0.296198, -0.171010, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.262003, -0.219846, 0.939693), vec3(-0.262003, -0.219846, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.433013, -0.250000, 0.866025), vec3(-0.433013, -0.250000, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.383022, -0.321394, 0.866025), vec3(-0.383022, -0.321394, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.556670, -0.321394, 0.766044), vec3(-0.556670, -0.321394, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, -0.413176, 0.766044), vec3(-0.492404, -0.413176, 0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.663414, -0.383022, 0.642788), vec3(-0.663414, -0.383022, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.586824, -0.492404, 0.642788), vec3(-0.586824, -0.492404, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.750000, -0.433013, 0.500000), vec3(-0.750000, -0.433013, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.663414, -0.556670, 0.500000), vec3(-0.663414, -0.556670, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.813798, -0.469846, 0.342020), vec3(-0.813798, -0.469846, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.719846, -0.604023, 0.342020), vec3(-0.719846, -0.604023, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.852869, -0.492404, 0.173648), vec3(-0.852869, -0.492404, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.754407, -0.633022, 0.173648), vec3(-0.754407, -0.633022, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.866025, -0.500000, 0.000000), vec3(-0.866025, -0.500000, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.766044, -0.642788, 0.000000), vec3(-0.766044, -0.642788, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.852869, -0.492404, -0.173648), vec3(-0.852869, -0.492404, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.754407, -0.633022, -0.173648), vec3(-0.754407, -0.633022, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.813798, -0.469846, -0.342020), vec3(-0.813798, -0.469846, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.719846, -0.604023, -0.342020), vec3(-0.719846, -0.604023, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.750000, -0.433013, -0.500000), vec3(-0.750000, -0.433013, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.663414, -0.556670, -0.500000), vec3(-0.663414, -0.556670, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.663414, -0.383022, -0.642788), vec3(-0.663414, -0.383022, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.586824, -0.492404, -0.642788), vec3(-0.586824, -0.492404, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.556670, -0.321394, -0.766044), vec3(-0.556670, -0.321394, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, -0.413176, -0.766044), vec3(-0.492404, -0.413176, -0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.433013, -0.250000, -0.866025), vec3(-0.433013, -0.250000, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.383022, -0.321394, -0.866025), vec3(-0.383022, -0.321394, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.296198, -0.171010, -0.939693), vec3(-0.296198, -0.171010, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.262003, -0.219846, -0.939693), vec3(-0.262003, -0.219846, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.150384, -0.086824, -0.984808), vec3(-0.150384, -0.086824, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.133022, -0.111619, -0.984808), vec3(-0.133022, -0.111619, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, -1.000000), vec3(0.000000, 0.000000, -1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.133022, -0.111619, -0.984808), vec3(-0.133022, -0.111619, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.111619, -0.133022, -0.984808), vec3(-0.111619, -0.133022, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.262003, -0.219846, -0.939693), vec3(-0.262003, -0.219846, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.219846, -0.262003, -0.939693), vec3(-0.219846, -0.262003, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.383022, -0.321394, -0.866025), vec3(-0.383022, -0.321394, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, -0.383022, -0.866025), vec3(-0.321394, -0.383022, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, -0.413176, -0.766044), vec3(-0.492404, -0.413176, -0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.413176, -0.492404, -0.766044), vec3(-0.413176, -0.492404, -0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.586824, -0.492404, -0.642788), vec3(-0.586824, -0.492404, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, -0.586824, -0.642788), vec3(-0.492404, -0.586824, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.663414, -0.556670, -0.500000), vec3(-0.663414, -0.556670, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.556670, -0.663414, -0.500000), vec3(-0.556670, -0.663414, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.719846, -0.604023, -0.342020), vec3(-0.719846, -0.604023, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.604023, -0.719846, -0.342020), vec3(-0.604023, -0.719846, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.754407, -0.633022, -0.173648), vec3(-0.754407, -0.633022, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.633022, -0.754407, -0.173648), vec3(-0.633022, -0.754407, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.766044, -0.642788, 0.000000), vec3(-0.766044, -0.642788, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.642788, -0.766044, 0.000000), vec3(-0.642788, -0.766044, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.754407, -0.633022, 0.173648), vec3(-0.754407, -0.633022, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.633022, -0.754407, 0.173648), vec3(-0.633022, -0.754407, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.719846, -0.604023, 0.342020), vec3(-0.719846, -0.604023, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.604023, -0.719846, 0.342020), vec3(-0.604023, -0.719846, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.663414, -0.556670, 0.500000), vec3(-0.663414, -0.556670, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.556670, -0.663414, 0.500000), vec3(-0.556670, -0.663414, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.586824, -0.492404, 0.642788), vec3(-0.586824, -0.492404, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, -0.586824, 0.642788), vec3(-0.492404, -0.586824, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, -0.413176, 0.766044), vec3(-0.492404, -0.413176, 0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.413176, -0.492404, 0.766044), vec3(-0.413176, -0.492404, 0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.383022, -0.321394, 0.866025), vec3(-0.383022, -0.321394, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, -0.383022, 0.866025), vec3(-0.321394, -0.383022, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.262003, -0.219846, 0.939693), vec3(-0.262003, -0.219846, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.219846, -0.262003, 0.939693), vec3(-0.219846, -0.262003, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.133022, -0.111619, 0.984808), vec3(-0.133022, -0.111619, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.111619, -0.133022, 0.984808), vec3(-0.111619, -0.133022, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, 1.000000), vec3(0.000000, 0.000000, 1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.111619, -0.133022, 0.984808), vec3(-0.111619, -0.133022, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.086824, -0.150384, 0.984808), vec3(-0.086824, -0.150384, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.219846, -0.262003, 0.939693), vec3(-0.219846, -0.262003, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, -0.296198, 0.939693), vec3(-0.171010, -0.296198, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, -0.383022, 0.866025), vec3(-0.321394, -0.383022, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.250000, -0.433013, 0.866025), vec3(-0.250000, -0.433013, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.413176, -0.492404, 0.766044), vec3(-0.413176, -0.492404, 0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, -0.556670, 0.766044), vec3(-0.321394, -0.556670, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, -0.586824, 0.642788), vec3(-0.492404, -0.586824, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.383022, -0.663414, 0.642788), vec3(-0.383022, -0.663414, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.556670, -0.663414, 0.500000), vec3(-0.556670, -0.663414, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.433013, -0.750000, 0.500000), vec3(-0.433013, -0.750000, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.604023, -0.719846, 0.342020), vec3(-0.604023, -0.719846, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.469846, -0.813798, 0.342020), vec3(-0.469846, -0.813798, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.633022, -0.754407, 0.173648), vec3(-0.633022, -0.754407, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, -0.852869, 0.173648), vec3(-0.492404, -0.852869, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.642788, -0.766044, 0.000000), vec3(-0.642788, -0.766044, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.500000, -0.866025, 0.000000), vec3(-0.500000, -0.866025, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.633022, -0.754407, -0.173648), vec3(-0.633022, -0.754407, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, -0.852869, -0.173648), vec3(-0.492404, -0.852869, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.604023, -0.719846, -0.342020), vec3(-0.604023, -0.719846, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.469846, -0.813798, -0.342020), vec3(-0.469846, -0.813798, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.556670, -0.663414, -0.500000), vec3(-0.556670, -0.663414, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.433013, -0.750000, -0.500000), vec3(-0.433013, -0.750000, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, -0.586824, -0.642788), vec3(-0.492404, -0.586824, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.383022, -0.663414, -0.642788), vec3(-0.383022, -0.663414, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.413176, -0.492404, -0.766044), vec3(-0.413176, -0.492404, -0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, -0.556670, -0.766044), vec3(-0.321394, -0.556670, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, -0.383022, -0.866025), vec3(-0.321394, -0.383022, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.250000, -0.433013, -0.866025), vec3(-0.250000, -0.433013, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.219846, -0.262003, -0.939693), vec3(-0.219846, -0.262003, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, -0.296198, -0.939693), vec3(-0.171010, -0.296198, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.111619, -0.133022, -0.984808), vec3(-0.111619, -0.133022, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.086824, -0.150384, -0.984808), vec3(-0.086824, -0.150384, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, -1.000000), vec3(0.000000, 0.000000, -1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.086824, -0.150384, -0.984808), vec3(-0.086824, -0.150384, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.059391, -0.163176, -0.984808), vec3(-0.059391, -0.163176, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, -0.296198, -0.939693), vec3(-0.171010, -0.296198, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.116978, -0.321394, -0.939693), vec3(-0.116978, -0.321394, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.250000, -0.433013, -0.866025), vec3(-0.250000, -0.433013, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, -0.469846, -0.866025), vec3(-0.171010, -0.469846, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, -0.556670, -0.766044), vec3(-0.321394, -0.556670, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.219846, -0.604023, -0.766044), vec3(-0.219846, -0.604023, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.383022, -0.663414, -0.642788), vec3(-0.383022, -0.663414, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.262003, -0.719846, -0.642788), vec3(-0.262003, -0.719846, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.433013, -0.750000, -0.500000), vec3(-0.433013, -0.750000, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.296198, -0.813798, -0.500000), vec3(-0.296198, -0.813798, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.469846, -0.813798, -0.342020), vec3(-0.469846, -0.813798, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, -0.883022, -0.342020), vec3(-0.321394, -0.883022, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, -0.852869, -0.173648), vec3(-0.492404, -0.852869, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.336824, -0.925417, -0.173648), vec3(-0.336824, -0.925417, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.500000, -0.866025, 0.000000), vec3(-0.500000, -0.866025, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.342020, -0.939693, 0.000000), vec3(-0.342020, -0.939693, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.492404, -0.852869, 0.173648), vec3(-0.492404, -0.852869, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.336824, -0.925417, 0.173648), vec3(-0.336824, -0.925417, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.469846, -0.813798, 0.342020), vec3(-0.469846, -0.813798, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, -0.883022, 0.342020), vec3(-0.321394, -0.883022, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.433013, -0.750000, 0.500000), vec3(-0.433013, -0.750000, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.296198, -0.813798, 0.500000), vec3(-0.296198, -0.813798, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.383022, -0.663414, 0.642788), vec3(-0.383022, -0.663414, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.262003, -0.719846, 0.642788), vec3(-0.262003, -0.719846, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, -0.556670, 0.766044), vec3(-0.321394, -0.556670, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.219846, -0.604023, 0.766044), vec3(-0.219846, -0.604023, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.250000, -0.433013, 0.866025), vec3(-0.250000, -0.433013, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, -0.469846, 0.866025), vec3(-0.171010, -0.469846, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, -0.296198, 0.939693), vec3(-0.171010, -0.296198, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.116978, -0.321394, 0.939693), vec3(-0.116978, -0.321394, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.086824, -0.150384, 0.984808), vec3(-0.086824, -0.150384, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.059391, -0.163176, 0.984808), vec3(-0.059391, -0.163176, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, 1.000000), vec3(0.000000, 0.000000, 1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.059391, -0.163176, 0.984808), vec3(-0.059391, -0.163176, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.030154, -0.171010, 0.984808), vec3(-0.030154, -0.171010, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.116978, -0.321394, 0.939693), vec3(-0.116978, -0.321394, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.059391, -0.336824, 0.939693), vec3(-0.059391, -0.336824, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, -0.469846, 0.866025), vec3(-0.171010, -0.469846, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.086824, -0.492404, 0.866025), vec3(-0.086824, -0.492404, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.219846, -0.604023, 0.766044), vec3(-0.219846, -0.604023, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.111619, -0.633022, 0.766044), vec3(-0.111619, -0.633022, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.262003, -0.719846, 0.642788), vec3(-0.262003, -0.719846, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.133022, -0.754407, 0.642788), vec3(-0.133022, -0.754407, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.296198, -0.813798, 0.500000), vec3(-0.296198, -0.813798, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.150384, -0.852869, 0.500000), vec3(-0.150384, -0.852869, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, -0.883022, 0.342020), vec3(-0.321394, -0.883022, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.163176, -0.925417, 0.342020), vec3(-0.163176, -0.925417, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.336824, -0.925417, 0.173648), vec3(-0.336824, -0.925417, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, -0.969846, 0.173648), vec3(-0.171010, -0.969846, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.342020, -0.939693, 0.000000), vec3(-0.342020, -0.939693, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.173648, -0.984808, 0.000000), vec3(-0.173648, -0.984808, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.336824, -0.925417, -0.173648), vec3(-0.336824, -0.925417, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, -0.969846, -0.173648), vec3(-0.171010, -0.969846, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.321394, -0.883022, -0.342020), vec3(-0.321394, -0.883022, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.163176, -0.925417, -0.342020), vec3(-0.163176, -0.925417, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.296198, -0.813798, -0.500000), vec3(-0.296198, -0.813798, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.150384, -0.852869, -0.500000), vec3(-0.150384, -0.852869, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.262003, -0.719846, -0.642788), vec3(-0.262003, -0.719846, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.133022, -0.754407, -0.642788), vec3(-0.133022, -0.754407, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.219846, -0.604023, -0.766044), vec3(-0.219846, -0.604023, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.111619, -0.633022, -0.766044), vec3(-0.111619, -0.633022, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, -0.469846, -0.866025), vec3(-0.171010, -0.469846, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.086824, -0.492404, -0.866025), vec3(-0.086824, -0.492404, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.116978, -0.321394, -0.939693), vec3(-0.116978, -0.321394, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.059391, -0.336824, -0.939693), vec3(-0.059391, -0.336824, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.059391, -0.163176, -0.984808), vec3(-0.059391, -0.163176, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.030154, -0.171010, -0.984808), vec3(-0.030154, -0.171010, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, -1.000000), vec3(0.000000, 0.000000, -1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.030154, -0.171010, -0.984808), vec3(-0.030154, -0.171010, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.173648, -0.984808), vec3(-0.000000, -0.173648, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.059391, -0.336824, -0.939693), vec3(-0.059391, -0.336824, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.342020, -0.939693), vec3(-0.000000, -0.342020, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.086824, -0.492404, -0.866025), vec3(-0.086824, -0.492404, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.500000, -0.866025), vec3(-0.000000, -0.500000, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.111619, -0.633022, -0.766044), vec3(-0.111619, -0.633022, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.642788, -0.766044), vec3(-0.000000, -0.642788, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.133022, -0.754407, -0.642788), vec3(-0.133022, -0.754407, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.766044, -0.642788), vec3(-0.000000, -0.766044, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.150384, -0.852869, -0.500000), vec3(-0.150384, -0.852869, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.866025, -0.500000), vec3(-0.000000, -0.866025, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.163176, -0.925417, -0.342020), vec3(-0.163176, -0.925417, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.939693, -0.342020), vec3(-0.000000, -0.939693, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, -0.969846, -0.173648), vec3(-0.171010, -0.969846, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.984808, -0.173648), vec3(-0.000000, -0.984808, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.173648, -0.984808, 0.000000), vec3(-0.173648, -0.984808, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -1.000000, 0.000000), vec3(-0.000000, -1.000000, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.171010, -0.969846, 0.173648), vec3(-0.171010, -0.969846, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.984808, 0.173648), vec3(-0.000000, -0.984808, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.163176, -0.925417, 0.342020), vec3(-0.163176, -0.925417, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.939693, 0.342020), vec3(-0.000000, -0.939693, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.150384, -0.852869, 0.500000), vec3(-0.150384, -0.852869, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.866025, 0.500000), vec3(-0.000000, -0.866025, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.133022, -0.754407, 0.642788), vec3(-0.133022, -0.754407, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.766044, 0.642788), vec3(-0.000000, -0.766044, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.111619, -0.633022, 0.766044), vec3(-0.111619, -0.633022, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.642788, 0.766044), vec3(-0.000000, -0.642788, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.086824, -0.492404, 0.866025), vec3(-0.086824, -0.492404, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.500000, 0.866025), vec3(-0.000000, -0.500000, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.059391, -0.336824, 0.939693), vec3(-0.059391, -0.336824, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.342020, 0.939693), vec3(-0.000000, -0.342020, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.030154, -0.171010, 0.984808), vec3(-0.030154, -0.171010, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.173648, 0.984808), vec3(-0.000000, -0.173648, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, 1.000000), vec3(0.000000, 0.000000, 1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.173648, 0.984808), vec3(-0.000000, -0.173648, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.030154, -0.171010, 0.984808), vec3(0.030154, -0.171010, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.342020, 0.939693), vec3(-0.000000, -0.342020, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.059391, -0.336824, 0.939693), vec3(0.059391, -0.336824, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.500000, 0.866025), vec3(-0.000000, -0.500000, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.086824, -0.492404, 0.866025), vec3(0.086824, -0.492404, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.642788, 0.766044), vec3(-0.000000, -0.642788, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.111619, -0.633022, 0.766044), vec3(0.111619, -0.633022, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.766044, 0.642788), vec3(-0.000000, -0.766044, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.133022, -0.754407, 0.642788), vec3(0.133022, -0.754407, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.866025, 0.500000), vec3(-0.000000, -0.866025, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.150384, -0.852869, 0.500000), vec3(0.150384, -0.852869, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.939693, 0.342020), vec3(-0.000000, -0.939693, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.163176, -0.925417, 0.342020), vec3(0.163176, -0.925417, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.984808, 0.173648), vec3(-0.000000, -0.984808, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, -0.969846, 0.173648), vec3(0.171010, -0.969846, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -1.000000, 0.000000), vec3(-0.000000, -1.000000, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.173648, -0.984808, 0.000000), vec3(0.173648, -0.984808, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.984808, -0.173648), vec3(-0.000000, -0.984808, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, -0.969846, -0.173648), vec3(0.171010, -0.969846, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.939693, -0.342020), vec3(-0.000000, -0.939693, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.163176, -0.925417, -0.342020), vec3(0.163176, -0.925417, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.866025, -0.500000), vec3(-0.000000, -0.866025, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.150384, -0.852869, -0.500000), vec3(0.150384, -0.852869, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.766044, -0.642788), vec3(-0.000000, -0.766044, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.133022, -0.754407, -0.642788), vec3(0.133022, -0.754407, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.642788, -0.766044), vec3(-0.000000, -0.642788, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.111619, -0.633022, -0.766044), vec3(0.111619, -0.633022, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.500000, -0.866025), vec3(-0.000000, -0.500000, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.086824, -0.492404, -0.866025), vec3(0.086824, -0.492404, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.342020, -0.939693), vec3(-0.000000, -0.342020, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.059391, -0.336824, -0.939693), vec3(0.059391, -0.336824, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(-0.000000, -0.173648, -0.984808), vec3(-0.000000, -0.173648, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.030154, -0.171010, -0.984808), vec3(0.030154, -0.171010, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, -1.000000), vec3(0.000000, 0.000000, -1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.030154, -0.171010, -0.984808), vec3(0.030154, -0.171010, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.059391, -0.163176, -0.984808), vec3(0.059391, -0.163176, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.059391, -0.336824, -0.939693), vec3(0.059391, -0.336824, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.116978, -0.321394, -0.939693), vec3(0.116978, -0.321394, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.086824, -0.492404, -0.866025), vec3(0.086824, -0.492404, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, -0.469846, -0.866025), vec3(0.171010, -0.469846, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.111619, -0.633022, -0.766044), vec3(0.111619, -0.633022, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.219846, -0.604023, -0.766044), vec3(0.219846, -0.604023, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.133022, -0.754407, -0.642788), vec3(0.133022, -0.754407, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.262003, -0.719846, -0.642788), vec3(0.262003, -0.719846, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.150384, -0.852869, -0.500000), vec3(0.150384, -0.852869, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.296198, -0.813798, -0.500000), vec3(0.296198, -0.813798, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.163176, -0.925417, -0.342020), vec3(0.163176, -0.925417, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, -0.883022, -0.342020), vec3(0.321394, -0.883022, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, -0.969846, -0.173648), vec3(0.171010, -0.969846, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.336824, -0.925417, -0.173648), vec3(0.336824, -0.925417, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.173648, -0.984808, 0.000000), vec3(0.173648, -0.984808, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.342020, -0.939693, 0.000000), vec3(0.342020, -0.939693, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, -0.969846, 0.173648), vec3(0.171010, -0.969846, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.336824, -0.925417, 0.173648), vec3(0.336824, -0.925417, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.163176, -0.925417, 0.342020), vec3(0.163176, -0.925417, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, -0.883022, 0.342020), vec3(0.321394, -0.883022, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.150384, -0.852869, 0.500000), vec3(0.150384, -0.852869, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.296198, -0.813798, 0.500000), vec3(0.296198, -0.813798, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.133022, -0.754407, 0.642788), vec3(0.133022, -0.754407, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.262003, -0.719846, 0.642788), vec3(0.262003, -0.719846, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.111619, -0.633022, 0.766044), vec3(0.111619, -0.633022, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.219846, -0.604023, 0.766044), vec3(0.219846, -0.604023, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.086824, -0.492404, 0.866025), vec3(0.086824, -0.492404, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, -0.469846, 0.866025), vec3(0.171010, -0.469846, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.059391, -0.336824, 0.939693), vec3(0.059391, -0.336824, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.116978, -0.321394, 0.939693), vec3(0.116978, -0.321394, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.030154, -0.171010, 0.984808), vec3(0.030154, -0.171010, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.059391, -0.163176, 0.984808), vec3(0.059391, -0.163176, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, 1.000000), vec3(0.000000, 0.000000, 1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.059391, -0.163176, 0.984808), vec3(0.059391, -0.163176, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.086824, -0.150384, 0.984808), vec3(0.086824, -0.150384, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.116978, -0.321394, 0.939693), vec3(0.116978, -0.321394, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, -0.296198, 0.939693), vec3(0.171010, -0.296198, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, -0.469846, 0.866025), vec3(0.171010, -0.469846, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.250000, -0.433013, 0.866025), vec3(0.250000, -0.433013, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.219846, -0.604023, 0.766044), vec3(0.219846, -0.604023, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, -0.556670, 0.766044), vec3(0.321394, -0.556670, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.262003, -0.719846, 0.642788), vec3(0.262003, -0.719846, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.383022, -0.663414, 0.642788), vec3(0.383022, -0.663414, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.296198, -0.813798, 0.500000), vec3(0.296198, -0.813798, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.433013, -0.750000, 0.500000), vec3(0.433013, -0.750000, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, -0.883022, 0.342020), vec3(0.321394, -0.883022, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.469846, -0.813798, 0.342020), vec3(0.469846, -0.813798, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.336824, -0.925417, 0.173648), vec3(0.336824, -0.925417, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, -0.852869, 0.173648), vec3(0.492404, -0.852869, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.342020, -0.939693, 0.000000), vec3(0.342020, -0.939693, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.500000, -0.866025, 0.000000), vec3(0.500000, -0.866025, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.336824, -0.925417, -0.173648), vec3(0.336824, -0.925417, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, -0.852869, -0.173648), vec3(0.492404, -0.852869, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, -0.883022, -0.342020), vec3(0.321394, -0.883022, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.469846, -0.813798, -0.342020), vec3(0.469846, -0.813798, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.296198, -0.813798, -0.500000), vec3(0.296198, -0.813798, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.433013, -0.750000, -0.500000), vec3(0.433013, -0.750000, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.262003, -0.719846, -0.642788), vec3(0.262003, -0.719846, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.383022, -0.663414, -0.642788), vec3(0.383022, -0.663414, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.219846, -0.604023, -0.766044), vec3(0.219846, -0.604023, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, -0.556670, -0.766044), vec3(0.321394, -0.556670, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, -0.469846, -0.866025), vec3(0.171010, -0.469846, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.250000, -0.433013, -0.866025), vec3(0.250000, -0.433013, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.116978, -0.321394, -0.939693), vec3(0.116978, -0.321394, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, -0.296198, -0.939693), vec3(0.171010, -0.296198, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.059391, -0.163176, -0.984808), vec3(0.059391, -0.163176, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.086824, -0.150384, -0.984808), vec3(0.086824, -0.150384, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, -1.000000), vec3(0.000000, 0.000000, -1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.086824, -0.150384, -0.984808), vec3(0.086824, -0.150384, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.111619, -0.133022, -0.984808), vec3(0.111619, -0.133022, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, -0.296198, -0.939693), vec3(0.171010, -0.296198, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.219846, -0.262003, -0.939693), vec3(0.219846, -0.262003, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.250000, -0.433013, -0.866025), vec3(0.250000, -0.433013, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, -0.383022, -0.866025), vec3(0.321394, -0.383022, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, -0.556670, -0.766044), vec3(0.321394, -0.556670, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.413176, -0.492404, -0.766044), vec3(0.413176, -0.492404, -0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.383022, -0.663414, -0.642788), vec3(0.383022, -0.663414, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, -0.586824, -0.642788), vec3(0.492404, -0.586824, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.433013, -0.750000, -0.500000), vec3(0.433013, -0.750000, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.556670, -0.663414, -0.500000), vec3(0.556670, -0.663414, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.469846, -0.813798, -0.342020), vec3(0.469846, -0.813798, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.604023, -0.719846, -0.342020), vec3(0.604023, -0.719846, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, -0.852869, -0.173648), vec3(0.492404, -0.852869, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.633022, -0.754407, -0.173648), vec3(0.633022, -0.754407, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.500000, -0.866025, 0.000000), vec3(0.500000, -0.866025, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.642788, -0.766044, 0.000000), vec3(0.642788, -0.766044, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, -0.852869, 0.173648), vec3(0.492404, -0.852869, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.633022, -0.754407, 0.173648), vec3(0.633022, -0.754407, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.469846, -0.813798, 0.342020), vec3(0.469846, -0.813798, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.604023, -0.719846, 0.342020), vec3(0.604023, -0.719846, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.433013, -0.750000, 0.500000), vec3(0.433013, -0.750000, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.556670, -0.663414, 0.500000), vec3(0.556670, -0.663414, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.383022, -0.663414, 0.642788), vec3(0.383022, -0.663414, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, -0.586824, 0.642788), vec3(0.492404, -0.586824, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, -0.556670, 0.766044), vec3(0.321394, -0.556670, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.413176, -0.492404, 0.766044), vec3(0.413176, -0.492404, 0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.250000, -0.433013, 0.866025), vec3(0.250000, -0.433013, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, -0.383022, 0.866025), vec3(0.321394, -0.383022, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, -0.296198, 0.939693), vec3(0.171010, -0.296198, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.219846, -0.262003, 0.939693), vec3(0.219846, -0.262003, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.086824, -0.150384, 0.984808), vec3(0.086824, -0.150384, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.111619, -0.133022, 0.984808), vec3(0.111619, -0.133022, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, 1.000000), vec3(0.000000, 0.000000, 1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.111619, -0.133022, 0.984808), vec3(0.111619, -0.133022, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.133022, -0.111619, 0.984808), vec3(0.133022, -0.111619, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.219846, -0.262003, 0.939693), vec3(0.219846, -0.262003, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.262003, -0.219846, 0.939693), vec3(0.262003, -0.219846, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, -0.383022, 0.866025), vec3(0.321394, -0.383022, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.383022, -0.321394, 0.866025), vec3(0.383022, -0.321394, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.413176, -0.492404, 0.766044), vec3(0.413176, -0.492404, 0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, -0.413176, 0.766044), vec3(0.492404, -0.413176, 0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, -0.586824, 0.642788), vec3(0.492404, -0.586824, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.586824, -0.492404, 0.642788), vec3(0.586824, -0.492404, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.556670, -0.663414, 0.500000), vec3(0.556670, -0.663414, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.663414, -0.556670, 0.500000), vec3(0.663414, -0.556670, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.604023, -0.719846, 0.342020), vec3(0.604023, -0.719846, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.719846, -0.604023, 0.342020), vec3(0.719846, -0.604023, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.633022, -0.754407, 0.173648), vec3(0.633022, -0.754407, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.754407, -0.633022, 0.173648), vec3(0.754407, -0.633022, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.642788, -0.766044, 0.000000), vec3(0.642788, -0.766044, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.766044, -0.642788, 0.000000), vec3(0.766044, -0.642788, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.633022, -0.754407, -0.173648), vec3(0.633022, -0.754407, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.754407, -0.633022, -0.173648), vec3(0.754407, -0.633022, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.604023, -0.719846, -0.342020), vec3(0.604023, -0.719846, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.719846, -0.604023, -0.342020), vec3(0.719846, -0.604023, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.556670, -0.663414, -0.500000), vec3(0.556670, -0.663414, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.663414, -0.556670, -0.500000), vec3(0.663414, -0.556670, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, -0.586824, -0.642788), vec3(0.492404, -0.586824, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.586824, -0.492404, -0.642788), vec3(0.586824, -0.492404, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.413176, -0.492404, -0.766044), vec3(0.413176, -0.492404, -0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, -0.413176, -0.766044), vec3(0.492404, -0.413176, -0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, -0.383022, -0.866025), vec3(0.321394, -0.383022, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.383022, -0.321394, -0.866025), vec3(0.383022, -0.321394, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.219846, -0.262003, -0.939693), vec3(0.219846, -0.262003, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.262003, -0.219846, -0.939693), vec3(0.262003, -0.219846, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.111619, -0.133022, -0.984808), vec3(0.111619, -0.133022, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.133022, -0.111619, -0.984808), vec3(0.133022, -0.111619, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, -1.000000), vec3(0.000000, 0.000000, -1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.133022, -0.111619, -0.984808), vec3(0.133022, -0.111619, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.150384, -0.086824, -0.984808), vec3(0.150384, -0.086824, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.262003, -0.219846, -0.939693), vec3(0.262003, -0.219846, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.296198, -0.171010, -0.939693), vec3(0.296198, -0.171010, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.383022, -0.321394, -0.866025), vec3(0.383022, -0.321394, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.433013, -0.250000, -0.866025), vec3(0.433013, -0.250000, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, -0.413176, -0.766044), vec3(0.492404, -0.413176, -0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.556670, -0.321394, -0.766044), vec3(0.556670, -0.321394, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.586824, -0.492404, -0.642788), vec3(0.586824, -0.492404, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.663414, -0.383022, -0.642788), vec3(0.663414, -0.383022, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.663414, -0.556670, -0.500000), vec3(0.663414, -0.556670, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.750000, -0.433013, -0.500000), vec3(0.750000, -0.433013, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.719846, -0.604023, -0.342020), vec3(0.719846, -0.604023, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.813798, -0.469846, -0.342020), vec3(0.813798, -0.469846, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.754407, -0.633022, -0.173648), vec3(0.754407, -0.633022, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.852869, -0.492404, -0.173648), vec3(0.852869, -0.492404, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.766044, -0.642788, 0.000000), vec3(0.766044, -0.642788, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.866025, -0.500000, 0.000000), vec3(0.866025, -0.500000, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.754407, -0.633022, 0.173648), vec3(0.754407, -0.633022, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.852869, -0.492404, 0.173648), vec3(0.852869, -0.492404, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.719846, -0.604023, 0.342020), vec3(0.719846, -0.604023, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.813798, -0.469846, 0.342020), vec3(0.813798, -0.469846, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.663414, -0.556670, 0.500000), vec3(0.663414, -0.556670, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.750000, -0.433013, 0.500000), vec3(0.750000, -0.433013, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.586824, -0.492404, 0.642788), vec3(0.586824, -0.492404, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.663414, -0.383022, 0.642788), vec3(0.663414, -0.383022, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, -0.413176, 0.766044), vec3(0.492404, -0.413176, 0.766045), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.556670, -0.321394, 0.766044), vec3(0.556670, -0.321394, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.383022, -0.321394, 0.866025), vec3(0.383022, -0.321394, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.433013, -0.250000, 0.866025), vec3(0.433013, -0.250000, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.262003, -0.219846, 0.939693), vec3(0.262003, -0.219846, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.296198, -0.171010, 0.939693), vec3(0.296198, -0.171010, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.133022, -0.111619, 0.984808), vec3(0.133022, -0.111619, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.150384, -0.086824, 0.984808), vec3(0.150384, -0.086824, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, 1.000000), vec3(0.000000, 0.000000, 1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.150384, -0.086824, 0.984808), vec3(0.150384, -0.086824, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.163176, -0.059391, 0.984808), vec3(0.163176, -0.059391, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.296198, -0.171010, 0.939693), vec3(0.296198, -0.171010, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, -0.116978, 0.939693), vec3(0.321394, -0.116978, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.433013, -0.250000, 0.866025), vec3(0.433013, -0.250000, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.469846, -0.171010, 0.866025), vec3(0.469846, -0.171010, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.556670, -0.321394, 0.766044), vec3(0.556670, -0.321394, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.604023, -0.219846, 0.766044), vec3(0.604023, -0.219846, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.663414, -0.383022, 0.642788), vec3(0.663414, -0.383022, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.719846, -0.262003, 0.642788), vec3(0.719846, -0.262003, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.750000, -0.433013, 0.500000), vec3(0.750000, -0.433013, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.813798, -0.296198, 0.500000), vec3(0.813798, -0.296198, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.813798, -0.469846, 0.342020), vec3(0.813798, -0.469846, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.883022, -0.321394, 0.342020), vec3(0.883022, -0.321394, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.852869, -0.492404, 0.173648), vec3(0.852869, -0.492404, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.925417, -0.336824, 0.173648), vec3(0.925417, -0.336824, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.866025, -0.500000, 0.000000), vec3(0.866025, -0.500000, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.939693, -0.342020, 0.000000), vec3(0.939693, -0.342020, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.852869, -0.492404, -0.173648), vec3(0.852869, -0.492404, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.925417, -0.336824, -0.173648), vec3(0.925417, -0.336824, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.813798, -0.469846, -0.342020), vec3(0.813798, -0.469846, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.883022, -0.321394, -0.342020), vec3(0.883022, -0.321394, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.750000, -0.433013, -0.500000), vec3(0.750000, -0.433013, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.813798, -0.296198, -0.500000), vec3(0.813798, -0.296198, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.663414, -0.383022, -0.642788), vec3(0.663414, -0.383022, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.719846, -0.262003, -0.642788), vec3(0.719846, -0.262003, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.556670, -0.321394, -0.766044), vec3(0.556670, -0.321394, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.604023, -0.219846, -0.766044), vec3(0.604023, -0.219846, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.433013, -0.250000, -0.866025), vec3(0.433013, -0.250000, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.469846, -0.171010, -0.866025), vec3(0.469846, -0.171010, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.296198, -0.171010, -0.939693), vec3(0.296198, -0.171010, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, -0.116978, -0.939693), vec3(0.321394, -0.116978, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.150384, -0.086824, -0.984808), vec3(0.150384, -0.086824, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.163176, -0.059391, -0.984808), vec3(0.163176, -0.059391, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, -1.000000), vec3(0.000000, 0.000000, -1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.163176, -0.059391, -0.984808), vec3(0.163176, -0.059391, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, -0.030154, -0.984808), vec3(0.171010, -0.030154, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, -0.116978, -0.939693), vec3(0.321394, -0.116978, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.336824, -0.059391, -0.939693), vec3(0.336824, -0.059391, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.469846, -0.171010, -0.866025), vec3(0.469846, -0.171010, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, -0.086824, -0.866025), vec3(0.492404, -0.086824, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.604023, -0.219846, -0.766044), vec3(0.604023, -0.219846, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.633022, -0.111619, -0.766044), vec3(0.633022, -0.111619, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.719846, -0.262003, -0.642788), vec3(0.719846, -0.262003, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.754407, -0.133022, -0.642788), vec3(0.754407, -0.133022, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.813798, -0.296198, -0.500000), vec3(0.813798, -0.296198, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.852869, -0.150384, -0.500000), vec3(0.852869, -0.150384, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.883022, -0.321394, -0.342020), vec3(0.883022, -0.321394, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.925417, -0.163176, -0.342020), vec3(0.925417, -0.163176, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.925417, -0.336824, -0.173648), vec3(0.925417, -0.336824, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.969846, -0.171010, -0.173648), vec3(0.969846, -0.171010, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.939693, -0.342020, 0.000000), vec3(0.939693, -0.342020, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.984808, -0.173648, 0.000000), vec3(0.984808, -0.173648, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.925417, -0.336824, 0.173648), vec3(0.925417, -0.336824, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.969846, -0.171010, 0.173648), vec3(0.969846, -0.171010, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.883022, -0.321394, 0.342020), vec3(0.883022, -0.321394, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.925417, -0.163176, 0.342020), vec3(0.925417, -0.163176, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.813798, -0.296198, 0.500000), vec3(0.813798, -0.296198, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.852869, -0.150384, 0.500000), vec3(0.852869, -0.150384, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.719846, -0.262003, 0.642788), vec3(0.719846, -0.262003, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.754407, -0.133022, 0.642788), vec3(0.754407, -0.133022, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.604023, -0.219846, 0.766044), vec3(0.604023, -0.219846, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.633022, -0.111619, 0.766044), vec3(0.633022, -0.111619, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.469846, -0.171010, 0.866025), vec3(0.469846, -0.171010, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, -0.086824, 0.866025), vec3(0.492404, -0.086824, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.321394, -0.116978, 0.939693), vec3(0.321394, -0.116978, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.336824, -0.059391, 0.939693), vec3(0.336824, -0.059391, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.163176, -0.059391, 0.984808), vec3(0.163176, -0.059391, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, -0.030154, 0.984808), vec3(0.171010, -0.030154, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, 1.000000), vec3(0.000000, 0.000000, 1.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, -0.030154, 0.984808), vec3(0.171010, -0.030154, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.173648, -0.000000, 0.984808), vec3(0.173648, -0.000000, 0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.336824, -0.059391, 0.939693), vec3(0.336824, -0.059391, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.342020, -0.000000, 0.939693), vec3(0.342020, -0.000000, 0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, -0.086824, 0.866025), vec3(0.492404, -0.086824, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.500000, -0.000000, 0.866025), vec3(0.500000, -0.000000, 0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.633022, -0.111619, 0.766044), vec3(0.633022, -0.111619, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.642788, -0.000000, 0.766044), vec3(0.642788, -0.000000, 0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.754407, -0.133022, 0.642788), vec3(0.754407, -0.133022, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.766044, -0.000000, 0.642788), vec3(0.766044, -0.000000, 0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.852869, -0.150384, 0.500000), vec3(0.852869, -0.150384, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.866025, -0.000000, 0.500000), vec3(0.866025, -0.000000, 0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.925417, -0.163176, 0.342020), vec3(0.925417, -0.163176, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.939693, -0.000000, 0.342020), vec3(0.939693, -0.000000, 0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.969846, -0.171010, 0.173648), vec3(0.969846, -0.171010, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.984808, -0.000000, 0.173648), vec3(0.984808, -0.000000, 0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.984808, -0.173648, 0.000000), vec3(0.984808, -0.173648, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(1.000000, -0.000000, 0.000000), vec3(1.000000, -0.000000, 0.000000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.969846, -0.171010, -0.173648), vec3(0.969846, -0.171010, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.984808, -0.000000, -0.173648), vec3(0.984808, -0.000000, -0.173648), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.925417, -0.163176, -0.342020), vec3(0.925417, -0.163176, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.939693, -0.000000, -0.342020), vec3(0.939693, -0.000000, -0.342020), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.852869, -0.150384, -0.500000), vec3(0.852869, -0.150384, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.866025, -0.000000, -0.500000), vec3(0.866025, -0.000000, -0.500000), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.754407, -0.133022, -0.642788), vec3(0.754407, -0.133022, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.766044, -0.000000, -0.642788), vec3(0.766044, -0.000000, -0.642788), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.633022, -0.111619, -0.766044), vec3(0.633022, -0.111619, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.642788, -0.000000, -0.766044), vec3(0.642788, -0.000000, -0.766044), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.492404, -0.086824, -0.866025), vec3(0.492404, -0.086824, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.500000, -0.000000, -0.866025), vec3(0.500000, -0.000000, -0.866025), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.336824, -0.059391, -0.939693), vec3(0.336824, -0.059391, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.342020, -0.000000, -0.939693), vec3(0.342020, -0.000000, -0.939693), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.171010, -0.030154, -0.984808), vec3(0.171010, -0.030154, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.173648, -0.000000, -0.984808), vec3(0.173648, -0.000000, -0.984808), vec3(1.0f, 0.05f, 0.05f) },
-        { vec3(0.000000, 0.000000, -1.000000), vec3(0.000000, 0.000000, -1.000000), vec3(1.0f, 0.05f, 0.05f) },
-    };
-
-    numOfVertices = sizeof(vertexBuffer) / sizeof(Vertex);
-
-    glGenVertexArrays(1, &mVAO);
-    glBindVertexArray(mVAO);
-    
-    glGenBuffers(1, &mVBO);
-    glBindBuffer(GL_ARRAY_BUFFER, mVBO);
-    glBufferData(GL_ARRAY_BUFFER, sizeof(vertexBuffer), vertexBuffer, GL_STATIC_DRAW);
-    
-    // 1st attribute buffer : vertex Positions
-    glVertexAttribPointer(  0,              // attribute. No particular reason for 0, but must match the layout in the shader.
-                          3,              // size
-                          GL_FLOAT,       // type
-                          GL_FALSE,       // normalized?
-                          sizeof(Vertex), // stride
-                          (void*)0        // array buffer offset
-                          );
-    glEnableVertexAttribArray(0);
-
-    // 2nd attribute buffer : vertex normal
-    glVertexAttribPointer(  1,
-                          3,
-                          GL_FLOAT,
-                          GL_FALSE,
-                          sizeof(Vertex),
-                          (void*)sizeof(vec3)    // Normal is Offseted by vec3 (see class Vertex)
-                          );
-    glEnableVertexAttribArray(1);
-
-    
-    // 3rd attribute buffer : vertex color
-    glVertexAttribPointer(  2,
-                          3,
-                          GL_FLOAT,
-                          GL_FALSE,
-                          sizeof(Vertex),
-                          (void*) (2* sizeof(vec3)) // Color is Offseted by 2 vec3 (see class Vertex)
-                          );
-    glEnableVertexAttribArray(2);
-}
-
-SphereModel::~SphereModel()
-{
-    glDeleteBuffers(1, &mVBO);
-    glDeleteVertexArrays(1, &mVAO);
-}
-
-
-void SphereModel::Update(float dt)
-{
-    Model::Update(dt);
-}
-
-void SphereModel::Draw(mat4 offsetMatrix)
-{
-    // Draw the Vertex Buffer
-    // Note this draws a Sphere
-    // The Model View Projection transforms are computed in the Vertex Shader
-    glBindVertexArray(mVAO);
-    glBindBuffer(GL_ARRAY_BUFFER, mVBO);
-
-    GLuint WorldMatrixLocation = glGetUniformLocation(Renderer::GetShaderProgramID(), "WorldTransform"); 
-	glm::mat4 wMatrix = offsetMatrix * GetWorldMatrix();
-    glUniformMatrix4fv(WorldMatrixLocation, 1, GL_FALSE, &wMatrix[0][0]);
-    
-
-	// Get a handle for Material Attributes uniform
-	GLuint MaterialID = glGetUniformLocation(Renderer::GetShaderProgramID(), "materialCoefficients");
-	// glUniform4f(MaterialID, ka, kd, ks, n);
-	glUniform4f(MaterialID, 0.2f, 0.8f, 0.2f, 50);
-
-    // Draw the triangles !
-    glDrawArrays(GL_TRIANGLE_STRIP, 0, numOfVertices);
-}
-
-bool SphereModel::ParseLine(const std::vector<ci_string> &token)
-{
-    if (token.empty())
-    {
-        return true;
-    }
-    else
-    {
-        return Model::ParseLine(token);
-    }
-}
