--- StaticCamera.cpp
@@ -1,42 +1,0 @@
-//
-// COMP 371 Assignment Framework
-//
-// Created by Nicolas Bergeron on 8/7/14.
-// Updated by Gary Chang on 14/1/15
-//
-// Copyright (c) 2014-2019 Concordia University. All rights reserved.
-//
-
-#include "StaticCamera.h"
-#include <glm/gtc/matrix_transform.hpp>
-#include <glm/gtx/string_cast.hpp>
-#include <iostream>
-
-#include "EventManager.h"
-
-using namespace glm;
-using namespace std;
-
-
-StaticCamera::StaticCamera(glm::vec3 position, glm::vec3 lookAtPoint, glm::vec3 upVector) 
-	: Camera(), mPosition(position), mLookAtPoint(lookAtPoint), mUpVector(upVector)
-{
-	printed = false;
-}
-
-StaticCamera::~StaticCamera()
-{
-}
-
-void StaticCamera::Update(float dt)
-{
-	EventManager::EnableMouseCursor();
-}
-
-glm::mat4 StaticCamera::GetViewMatrix() const
-{
-	return glm::lookAt(	mPosition,		// Camera position
-                        mLookAtPoint,	// Look towards this point
-                        mUpVector		// Up vector
-						);
-}
